Move Unstable module to the main parquet folder; Remove Unstable Module
