+++ dataframe.cabal
@@ -80,28 +80,18 @@ library
                     DataFrame.Display.Terminal.Plot,
                     DataFrame.IO.CSV,
                     DataFrame.IO.JSON,
-                    DataFrame.IO.Unstable.Parquet.Utils,
-                    DataFrame.IO.Unstable.Parquet.Encoding,
-                    DataFrame.IO.Unstable.Parquet.Levels,
-                    DataFrame.IO.Unstable.Parquet.Dictionary,
-                    DataFrame.IO.Unstable.Parquet.Time,
-                    DataFrame.IO.Unstable.Parquet.Thrift,
-                    DataFrame.IO.Unstable.Parquet.Decompress,
-                    DataFrame.IO.Unstable.Parquet.Page,
-                    DataFrame.IO.Unstable.Parquet,
                     DataFrame.IO.Utils.RandomAccess,
                     DataFrame.IO.Parquet,
                     DataFrame.IO.Parquet.Binary,
                     DataFrame.IO.Parquet.Dictionary,
                     DataFrame.IO.Parquet.Levels,
                     DataFrame.IO.Parquet.Thrift,
-                    DataFrame.IO.Parquet.ColumnStatistics,
-                    DataFrame.IO.Parquet.Compression,
+                    DataFrame.IO.Parquet.Decompress,
                     DataFrame.IO.Parquet.Encoding,
                     DataFrame.IO.Parquet.Page,
+                    DataFrame.IO.Parquet.Utils,
                     DataFrame.IO.Parquet.Seeking,
                     DataFrame.IO.Parquet.Time,
-                    DataFrame.IO.Parquet.Types,
                     DataFrame.Lazy.IO.CSV,
                     DataFrame.Lazy.IO.Binary,
                     DataFrame.Lazy.Internal.DataFrame,
