+++ examples/examples.cabal
@@ -61,6 +61,7 @@ executable examples
                    DataFrame.IO.JSON,
                    DataFrame.IO.Parquet,
                    DataFrame.IO.Parquet.Binary,
+                   DataFrame.IO.Parquet.Decompress,
                    DataFrame.IO.Parquet.Dictionary,
                    DataFrame.IO.Parquet.Levels,
                    DataFrame.IO.Parquet.Thrift,
@@ -70,6 +71,8 @@ executable examples
                    DataFrame.IO.Parquet.Page,
                    DataFrame.IO.Parquet.Time,
                    DataFrame.IO.Parquet.Types,
+                   DataFrame.IO.Parquet.Utils,
+                   DataFrame.IO.Utils.RandomAccess,
                    DataFrame.Lazy.IO.CSV,
                    DataFrame.Lazy.IO.Binary,
                    DataFrame.Lazy.Internal.DataFrame,
