+++ src/DataFrame/IO/Parquet/Encoding.hs
@@ -1,8 +1,18 @@
 {-# LANGUAGE BangPatterns #-}
 {-# LANGUAGE CPP #-}
 
-module DataFrame.IO.Parquet.Encoding where
+module DataFrame.IO.Parquet.Encoding (
+    -- Kept from the original Encoding module (used by Levels)
+    ceilLog2,
+    bitWidthForMaxLevel,
+    -- Vector-based RLE/bit-packed decoder (from new parser)
+    decodeRLEBitPackedHybridV,
+    extractBitsIntoV,
+    fillRun,
+    decodeDictIndicesV,
+) where
 
+import Control.Monad.ST (ST, runST)
 import Data.Bits
 import qualified Data.ByteString as BS
 import qualified Data.ByteString.Unsafe as BSU
@@ -10,9 +20,14 @@ import qualified Data.ByteString.Unsafe as BSU
 import Data.List (foldl')
 #endif
 import qualified Data.Vector.Unboxed as VU
+import qualified Data.Vector.Unboxed.Mutable as VUM
 import Data.Word
 import DataFrame.IO.Parquet.Binary (readUVarInt)
 import DataFrame.Internal.Binary (littleEndianWord32)
+
+-- ---------------------------------------------------------------------------
+-- Level-width helpers (used by Levels.hs)
+-- ---------------------------------------------------------------------------
 
 ceilLog2 :: Int -> Int
 ceilLog2 x
@@ -22,73 +37,101 @@ ceilLog2 x
 bitWidthForMaxLevel :: Int -> Int
 bitWidthForMaxLevel maxLevel = ceilLog2 (maxLevel + 1)
 
-bytesForBW :: Int -> Int
-bytesForBW bw = (bw + 7) `div` 8
+-- ---------------------------------------------------------------------------
+-- Vector-based RLE / bit-packed hybrid decoder
+-- ---------------------------------------------------------------------------
 
-unpackBitPacked :: Int -> Int -> BS.ByteString -> ([Word32], BS.ByteString)
-unpackBitPacked bw count bs
-    | count <= 0 = ([], bs)
-    | BS.null bs = ([], bs)
-    | otherwise =
-        let totalBytes = (bw * count + 7) `div` 8
-            chunk = BS.take totalBytes bs
-            rest = BS.drop totalBytes bs
-         in (extractBits bw count chunk, rest)
+decodeRLEBitPackedHybridV ::
+    -- | Bit width per value (0 = all zeros, use 'VU.replicate')
+    Int ->
+    -- | Exact number of values to decode
+    Int ->
+    BS.ByteString ->
+    (VU.Vector Word32, BS.ByteString)
+decodeRLEBitPackedHybridV bw need bs
+    | bw == 0 = (VU.replicate need 0, bs)
+    | otherwise = runST $ do
+        mv <- VUM.new need
+        rest <- go mv 0 bs
+        dat <- VU.unsafeFreeze mv
+        return (dat, rest)
+  where
+    !mask = if bw == 32 then maxBound else (1 `shiftL` bw) - 1 :: Word32
+    go :: VUM.STVector s Word32 -> Int -> BS.ByteString -> ST s BS.ByteString
+    go mv !filled !buf
+        | filled >= need = return buf
+        | BS.null buf = return buf
+        | otherwise =
+            let (hdr64, afterHdr) = readUVarInt buf
+                isPacked = (hdr64 .&. 1) == 1
+             in if isPacked
+                    then do
+                        let groups = fromIntegral (hdr64 `shiftR` 1) :: Int
+                            totalVals = groups * 8
+                            takeN = min (need - filled) totalVals
+                            -- Consume all the bytes for this group even if we
+                            -- only need a subset of the values.
+                            bytesN = (bw * totalVals + 7) `div` 8
+                            (chunk, rest) = BS.splitAt bytesN afterHdr
+                        extractBitsIntoV bw takeN chunk mv filled
+                        go mv (filled + takeN) rest
+                    else do
+                        let runLen = fromIntegral (hdr64 `shiftR` 1) :: Int
+                            nbytes = (bw + 7) `div` 8
+                            val = littleEndianWord32 (BS.take 4 afterHdr) .&. mask
+                            takeN = min (need - filled) runLen
+                        -- Fill the run directly — no list, no reverse.
+                        fillRun mv filled (filled + takeN) val
+                        go mv (filled + takeN) (BS.drop nbytes afterHdr)
+{-# INLINE decodeRLEBitPackedHybridV #-}
 
--- | LSB-first bit accumulator: reads each byte once with no intermediate ByteString allocation.
-extractBits :: Int -> Int -> BS.ByteString -> [Word32]
-extractBits bw count bs = go 0 (0 :: Word64) 0 count
+-- | Fill @mv[start..end-1]@ with @val@.
+fillRun :: VUM.STVector s Word32 -> Int -> Int -> Word32 -> ST s ()
+fillRun mv !i !end !val
+    | i >= end = return ()
+    | otherwise = VUM.unsafeWrite mv i val >> fillRun mv (i + 1) end val
+{-# INLINE fillRun #-}
+
+{- | Write @count@ bit-width-@bw@ values from @bs@ into @mv@ starting at
+@offset@, reading the byte buffer with a single-pass LSB-first accumulator.
+No intermediate list or ByteString allocation.
+-}
+extractBitsIntoV ::
+    -- | Bit width
+    Int ->
+    -- | Number of values to extract
+    Int ->
+    BS.ByteString ->
+    VUM.STVector s Word32 ->
+    -- | Write offset into @mv@
+    Int ->
+    ST s ()
+extractBitsIntoV bw count bs mv off = go 0 (0 :: Word64) 0 0
   where
     !mask = if bw == 32 then maxBound else (1 `shiftL` bw) - 1 :: Word64
     !len = BS.length bs
-    go !byteIdx !acc !accBits !remaining
-        | remaining <= 0 = []
-        | accBits >= bw =
-            fromIntegral (acc .&. mask)
-                : go byteIdx (acc `shiftR` bw) (accBits - bw) (remaining - 1)
-        | byteIdx >= len = []
+    go !byteIdx !acc !accBits !done
+        | done >= count = return ()
+        | accBits >= bw = do
+            VUM.unsafeWrite mv (off + done) (fromIntegral (acc .&. mask))
+            go byteIdx (acc `shiftR` bw) (accBits - bw) (done + 1)
+        | byteIdx >= len = return ()
         | otherwise =
             let b = fromIntegral (BSU.unsafeIndex bs byteIdx) :: Word64
-             in go (byteIdx + 1) (acc .|. (b `shiftL` accBits)) (accBits + 8) remaining
+             in go (byteIdx + 1) (acc .|. (b `shiftL` accBits)) (accBits + 8) done
+{-# INLINE extractBitsIntoV #-}
 
-decodeRLEBitPackedHybrid ::
-    Int -> Int -> BS.ByteString -> ([Word32], BS.ByteString)
-decodeRLEBitPackedHybrid bw need bs
-    | bw == 0 = (replicate need 0, bs)
-    | otherwise = go need bs []
-  where
-    mask :: Word32
-    mask = if bw == 32 then maxBound else (1 `shiftL` bw) - 1
-    go :: Int -> BS.ByteString -> [Word32] -> ([Word32], BS.ByteString)
-    go 0 rest acc = (reverse acc, rest)
-    go n rest acc
-        | BS.null rest = (reverse acc, rest)
-        | otherwise =
-            let (hdr64, afterHdr) = readUVarInt rest
-                isPacked = (hdr64 .&. 1) == 1
-             in if isPacked
-                    then
-                        let groups = fromIntegral (hdr64 `shiftR` 1) :: Int
-                            totalVals = groups * 8
-                            (valsAll, afterRun) = unpackBitPacked bw totalVals afterHdr
-                            takeN = min n totalVals
-                            actualTaken = take takeN valsAll
-                         in go (n - takeN) afterRun (reverse actualTaken ++ acc)
-                    else
-                        let runLen = fromIntegral (hdr64 `shiftR` 1) :: Int
-                            nbytes = bytesForBW bw
-                            word32 = littleEndianWord32 (BS.take 4 afterHdr)
-                            afterV = BS.drop nbytes afterHdr
-                            val = word32 .&. mask
-                            takeN = min n runLen
-                         in go (n - takeN) afterV (replicate takeN val ++ acc)
+{- | Decode @need@ dictionary indices from a DATA_PAGE bit-width-prefixed
+stream (the first byte encodes the bit-width of all subsequent RLE\/bitpacked
+values).
 
-decodeDictIndicesV1 ::
-    Int -> Int -> BS.ByteString -> (VU.Vector Int, BS.ByteString)
-decodeDictIndicesV1 need _dictCard bs =
-    case BS.uncons bs of
-        Nothing -> error "empty dictionary index stream"
-        Just (w0, rest0) ->
-            let bw = fromIntegral w0 :: Int
-                (u32s, rest1) = decodeRLEBitPackedHybrid bw need rest0
-             in (VU.fromList (map fromIntegral u32s), rest1)
+Returns the index vector (as 'Int') and the unconsumed bytes.
+-}
+decodeDictIndicesV :: Int -> BS.ByteString -> (VU.Vector Int, BS.ByteString)
+decodeDictIndicesV need bs = case BS.uncons bs of
+    Nothing -> error "decodeDictIndicesV: empty stream"
+    Just (w0, rest0) ->
+        let bw = fromIntegral w0 :: Int
+            (raw, rest1) = decodeRLEBitPackedHybridV bw need rest0
+         in (VU.map fromIntegral raw, rest1)
+{-# INLINE decodeDictIndicesV #-}
