+++ src/DataFrame/IO/Parquet/Levels.hs
@@ -1,145 +1,145 @@
-module DataFrame.IO.Parquet.Levels where
+module DataFrame.IO.Parquet.Levels (
+    -- Level readers
+    readLevelsV1V,
+    readLevelsV2V,
+    -- Stitch functions
+    stitchNullableV,
+    stitchListV,
+    stitchList2V,
+    stitchList3V,
+) where
 
+import Control.Monad.ST (runST)
 import qualified Data.ByteString as BS
-import Data.Int
-import Data.List
-import qualified Data.Text as T
-
-import DataFrame.IO.Parquet.Encoding
-import DataFrame.IO.Parquet.Thrift
-import DataFrame.IO.Parquet.Types
+import Data.Int (Int32)
+import qualified Data.Vector as VB
+import qualified Data.Vector.Mutable as VBM
+import qualified Data.Vector.Unboxed as VU
+import Data.Word (Word32)
+import DataFrame.IO.Parquet.Encoding (
+    bitWidthForMaxLevel,
+    decodeRLEBitPackedHybridV,
+ )
 import DataFrame.Internal.Binary (littleEndianWord32)
 
-readLevelsV1 ::
-    Int -> Int -> Int -> BS.ByteString -> ([Int], [Int], BS.ByteString)
-readLevelsV1 n maxDef maxRep bs =
-    let bwDef = bitWidthForMaxLevel maxDef
-        bwRep = bitWidthForMaxLevel maxRep
+-- ---------------------------------------------------------------------------
+-- Level readers
+-- ---------------------------------------------------------------------------
 
-        (repLvls, afterRep) =
-            if bwRep == 0
-                then (replicate n 0, bs)
-                else
-                    let repLength = littleEndianWord32 (BS.take 4 bs)
-                        repData = BS.take (fromIntegral repLength) (BS.drop 4 bs)
-                        afterRepData = BS.drop (4 + fromIntegral repLength) bs
-                        (repVals, _) = decodeRLEBitPackedHybrid bwRep n repData
-                     in (map fromIntegral repVals, afterRepData)
+readLevelsV1V ::
+    -- | Total number of values in the page
+    Int ->
+    -- | maxDefinitionLevel
+    Int ->
+    -- | maxRepetitionLevel
+    Int ->
+    BS.ByteString ->
+    (VU.Vector Int, VU.Vector Int, Int, BS.ByteString)
+readLevelsV1V n maxDef maxRep bs =
+    let bwRep = bitWidthForMaxLevel maxRep
+        bwDef = bitWidthForMaxLevel maxDef
+        (repVec, afterRep) = decodeLevelBlock bwRep n bs
+        (defVec, afterDef) = decodeLevelBlock bwDef n afterRep
+        nPresent = VU.foldl' (\acc d -> acc + fromEnum (d == maxDef)) 0 defVec
+     in (defVec, repVec, nPresent, afterDef)
+  where
+    decodeLevelBlock 0 n' buf = (VU.replicate n' 0, buf)
+    decodeLevelBlock bw n' buf =
+        let blockLen = fromIntegral (littleEndianWord32 (BS.take 4 buf)) :: Int
+            blockData = BS.take blockLen (BS.drop 4 buf)
+            after = BS.drop (4 + blockLen) buf
+            (raw, _) = decodeRLEBitPackedHybridV bw n' blockData
+         in (VU.map (fromIntegral :: Word32 -> Int) raw, after)
 
-        (defLvls, afterDef) =
-            if bwDef == 0
-                then (replicate n 0, afterRep)
-                else
-                    let defLength = littleEndianWord32 (BS.take 4 afterRep)
-                        defData = BS.take (fromIntegral defLength) (BS.drop 4 afterRep)
-                        afterDefData = BS.drop (4 + fromIntegral defLength) afterRep
-                        (defVals, _) = decodeRLEBitPackedHybrid bwDef n defData
-                     in (map fromIntegral defVals, afterDefData)
-     in (defLvls, repLvls, afterDef)
-
-readLevelsV2 ::
+readLevelsV2V ::
+    -- | Total number of values
     Int ->
+    -- | maxDefinitionLevel
     Int ->
+    -- | maxRepetitionLevel
     Int ->
+    -- | Repetition-level byte length (from page header)
     Int32 ->
+    -- | Definition-level byte length (from page header)
     Int32 ->
     BS.ByteString ->
-    ([Int], [Int], BS.ByteString)
-readLevelsV2 n maxDef maxRep defLen repLen bs =
+    (VU.Vector Int, VU.Vector Int, Int, BS.ByteString)
+readLevelsV2V n maxDef maxRep repLen defLen bs =
     let (repBytes, afterRepBytes) = BS.splitAt (fromIntegral repLen) bs
         (defBytes, afterDefBytes) = BS.splitAt (fromIntegral defLen) afterRepBytes
-        bwDef = bitWidthForMaxLevel maxDef
         bwRep = bitWidthForMaxLevel maxRep
-        (repLvlsRaw, _) =
-            if bwRep == 0
-                then (replicate n 0, repBytes)
-                else decodeRLEBitPackedHybrid bwRep n repBytes
-        (defLvlsRaw, _) =
-            if bwDef == 0
-                then (replicate n 0, defBytes)
-                else decodeRLEBitPackedHybrid bwDef n defBytes
-     in (map fromIntegral defLvlsRaw, map fromIntegral repLvlsRaw, afterDefBytes)
+        bwDef = bitWidthForMaxLevel maxDef
+        repVec
+            | bwRep == 0 = VU.replicate n 0
+            | otherwise =
+                let (raw, _) = decodeRLEBitPackedHybridV bwRep n repBytes
+                 in VU.map (fromIntegral :: Word32 -> Int) raw
+        defVec
+            | bwDef == 0 = VU.replicate n 0
+            | otherwise =
+                let (raw, _) = decodeRLEBitPackedHybridV bwDef n defBytes
+                 in VU.map (fromIntegral :: Word32 -> Int) raw
+        nPresent = VU.foldl' (\acc d -> acc + fromEnum (d == maxDef)) 0 defVec
+     in (defVec, repVec, nPresent, afterDefBytes)
 
-stitchNullable :: Int -> [Int] -> [a] -> [Maybe a]
-stitchNullable maxDef = go
-  where
-    go [] _ = []
-    go (d : ds) vs
-        | d == maxDef = case vs of
-            (v : vs') -> Just v : go ds vs'
-            [] -> error "value stream exhausted"
-        | otherwise = Nothing : go ds vs
+{- | Build a full-length vector of @Maybe a@ from definition levels and a
+compact present-values vector.
 
-data SNode = SNode
-    { sName :: String
-    , sRep :: RepetitionType
-    , sChildren :: [SNode]
-    }
-    deriving (Show, Eq)
+For each index @i@:
 
-parseOne :: [SchemaElement] -> (SNode, [SchemaElement])
-parseOne [] = error "parseOne: empty schema list"
-parseOne (se : rest) =
-    let childCount = fromIntegral (numChildren se)
-        (kids, rest') = parseMany childCount rest
-     in ( SNode
-            { sName = T.unpack (elementName se)
-            , sRep = repetitionType se
-            , sChildren = kids
-            }
-        , rest'
-        )
+  * @defVec VU.! i == maxDef@  →  @Just (values VB.! j)@, advancing @j@
+  * @defVec VU.! i <  maxDef@  →  @Nothing@
 
-parseMany :: Int -> [SchemaElement] -> ([SNode], [SchemaElement])
-parseMany 0 xs = ([], xs)
-parseMany n xs =
-    let (node, xs') = parseOne xs
-        (nodes, xs'') = parseMany (n - 1) xs'
-     in (node : nodes, xs'')
-
-parseAll :: [SchemaElement] -> [SNode]
-parseAll [] = []
-parseAll xs = let (n, xs') = parseOne xs in n : parseAll xs'
-
--- | Tag leaf values as Just/Nothing according to maxDef.
-pairWithVals :: Int -> [(Int, Int)] -> [a] -> [(Int, Int, Maybe a)]
-pairWithVals _ [] _ = []
-pairWithVals maxDef ((r, d) : rds) vs
-    | d == maxDef = case vs of
-        (v : vs') -> (r, d, Just v) : pairWithVals maxDef rds vs'
-        [] -> error "pairWithVals: value stream exhausted"
-    | otherwise = (r, d, Nothing) : pairWithVals maxDef rds vs
-
--- | Split triplets into groups; a new group begins whenever rep <= bound.
-splitAtRepBound :: Int -> [(Int, Int, Maybe a)] -> [[(Int, Int, Maybe a)]]
-splitAtRepBound _ [] = []
-splitAtRepBound bound (t : ts) =
-    let (rest, remaining) = span (\(r, _, _) -> r > bound) ts
-     in (t : rest) : splitAtRepBound bound remaining
-
-{- | Reconstruct a list column from Dremel encoding levels.
-rep=0 starts a new top-level row; def=0 means the entire list slot is null.
-Returns one Maybe [Maybe a] per row.
+The length of the result equals @VU.length defVec@.
 -}
-stitchList :: Int -> [Int] -> [Int] -> [a] -> [Maybe [Maybe a]]
-stitchList maxDef repLvls defLvls vals =
-    let triplets = pairWithVals maxDef (zip repLvls defLvls) vals
-        rows = splitAtRepBound 0 triplets
-     in map toRow rows
+stitchNullableV ::
+    Int ->
+    VU.Vector Int ->
+    VB.Vector a ->
+    VB.Vector (Maybe a)
+stitchNullableV maxDef defVec values = runST $ do
+    let n = VU.length defVec
+    mv <- VBM.replicate n Nothing
+    let go i j
+            | i >= n = pure ()
+            | VU.unsafeIndex defVec i == maxDef = do
+                VBM.unsafeWrite mv i (Just (VB.unsafeIndex values j))
+                go (i + 1) (j + 1)
+            | otherwise = go (i + 1) j
+    go 0 0
+    VB.unsafeFreeze mv
+
+{- | Stitch a singly-nested list column (@maxRep == 1@) from vector-format
+definition and repetition levels plus a compact present-values vector.
+Returns one @Maybe [Maybe a]@ per top-level row.
+-}
+stitchListV ::
+    Int ->
+    VU.Vector Int ->
+    VU.Vector Int ->
+    VB.Vector a ->
+    [Maybe [Maybe a]]
+stitchListV maxDef repVec defVec values =
+    map toRow (splitAtRepBound 0 (pairWithValsV maxDef repVec defVec values))
   where
     toRow [] = Nothing
     toRow ((_, d, _) : _) | d == 0 = Nothing
     toRow grp = Just [v | (_, _, v) <- grp]
 
-{- | Reconstruct a 2-level nested list (maxRep=2) from Dremel triplets.
-defT1: def threshold at which the depth-1 element is present (not null).
-maxDef: def threshold at which the leaf is present.
+{- | Stitch a doubly-nested list column (@maxRep == 2@).
+@defT1@ is the def threshold at which the depth-1 element is present.
 -}
-stitchList2 :: Int -> Int -> [Int] -> [Int] -> [a] -> [Maybe [Maybe [Maybe a]]]
-stitchList2 defT1 maxDef repLvls defLvls vals =
-    let triplets = pairWithVals maxDef (zip repLvls defLvls) vals
-     in map toRow (splitAtRepBound 0 triplets)
+stitchList2V ::
+    Int ->
+    Int ->
+    VU.Vector Int ->
+    VU.Vector Int ->
+    VB.Vector a ->
+    [Maybe [Maybe [Maybe a]]]
+stitchList2V defT1 maxDef repVec defVec values =
+    map toRow (splitAtRepBound 0 triplets)
   where
+    triplets = pairWithValsV maxDef repVec defVec values
     toRow [] = Nothing
     toRow ((_, d, _) : _) | d == 0 = Nothing
     toRow row = Just (map toOuter (splitAtRepBound 1 row))
@@ -149,16 +149,22 @@ stitchList2 defT1 maxDef repLvls defLvls vals =
     toLeaf [] = Nothing
     toLeaf ((_, _, v) : _) = v
 
-{- | Reconstruct a 3-level nested list (maxRep=3) from Dremel triplets.
-defT1, defT2: def thresholds at which depth-1 and depth-2 elements are present.
-maxDef: def threshold at which the leaf is present.
+{- | Stitch a triply-nested list column (@maxRep == 3@).
+@defT1@ and @defT2@ are the def thresholds for depth-1 and depth-2
+elements respectively.
 -}
-stitchList3 ::
-    Int -> Int -> Int -> [Int] -> [Int] -> [a] -> [Maybe [Maybe [Maybe [Maybe a]]]]
-stitchList3 defT1 defT2 maxDef repLvls defLvls vals =
-    let triplets = pairWithVals maxDef (zip repLvls defLvls) vals
-     in map toRow (splitAtRepBound 0 triplets)
+stitchList3V ::
+    Int ->
+    Int ->
+    Int ->
+    VU.Vector Int ->
+    VU.Vector Int ->
+    VB.Vector a ->
+    [Maybe [Maybe [Maybe [Maybe a]]]]
+stitchList3V defT1 defT2 maxDef repVec defVec values =
+    map toRow (splitAtRepBound 0 triplets)
   where
+    triplets = pairWithValsV maxDef repVec defVec values
     toRow [] = Nothing
     toRow ((_, d, _) : _) | d == 0 = Nothing
     toRow row = Just (map toOuter (splitAtRepBound 1 row))
@@ -171,14 +177,37 @@ stitchList3 defT1 defT2 maxDef repLvls defLvls vals =
     toLeaf [] = Nothing
     toLeaf ((_, _, v) : _) = v
 
-levelsForPath :: [SchemaElement] -> [String] -> (Int, Int)
-levelsForPath schemaTail = go 0 0 (parseAll schemaTail)
+-- ---------------------------------------------------------------------------
+-- Internal helpers
+-- ---------------------------------------------------------------------------
+
+{- | Zip rep and def level vectors with a present-values vector, tagging each
+position as @Just value@ (when @def == maxDef@) or @Nothing@.
+Returns a flat list of @(rep, def, Maybe a)@ triplets for row-splitting.
+-}
+pairWithValsV ::
+    Int ->
+    VU.Vector Int ->
+    VU.Vector Int ->
+    VB.Vector a ->
+    [(Int, Int, Maybe a)]
+pairWithValsV maxDef repVec defVec values = go 0 0
   where
-    go defC repC _ [] = (defC, repC)
-    go defC repC nodes (p : ps) =
-        case find (\n -> sName n == p) nodes of
-            Nothing -> (defC, repC)
-            Just n ->
-                let defC' = defC + (if sRep n == OPTIONAL || sRep n == REPEATED then 1 else 0)
-                    repC' = repC + (if sRep n == REPEATED then 1 else 0)
-                 in go defC' repC' (sChildren n) ps
+    n = VU.length defVec
+    go i j
+        | i >= n = []
+        | otherwise =
+            let r = VU.unsafeIndex repVec i
+                d = VU.unsafeIndex defVec i
+             in if d == maxDef
+                    then (r, d, Just (VB.unsafeIndex values j)) : go (i + 1) (j + 1)
+                    else (r, d, Nothing) : go (i + 1) j
+
+{- | Group a flat triplet list into rows.
+A new group begins whenever @rep <= bound@.
+-}
+splitAtRepBound :: Int -> [(Int, Int, Maybe a)] -> [[(Int, Int, Maybe a)]]
+splitAtRepBound _ [] = []
+splitAtRepBound bound (t : ts) =
+    let (rest, remaining) = span (\(r, _, _) -> r > bound) ts
+     in (t : rest) : splitAtRepBound bound remaining
