+++ src/DataFrame/IO/Parquet/Page.hs
@@ -1,473 +1,334 @@
-{-# LANGUAGE TypeApplications #-}
+{-# LANGUAGE OverloadedRecordDot #-}
+{-# LANGUAGE ScopedTypeVariables #-}
 
-module DataFrame.IO.Parquet.Page where
+module DataFrame.IO.Parquet.Page (
+    -- Types
+    PageDecoder,
+    -- Per-type decoders
+    boolDecoder,
+    int32Decoder,
+    int64Decoder,
+    int96Decoder,
+    floatDecoder,
+    doubleDecoder,
+    byteArrayDecoder,
+    fixedLenByteArrayDecoder,
+    -- Page iteration
+    readPages,
+) where
 
-import qualified Codec.Compression.GZip as GZip
-import qualified Codec.Compression.Zstd.Streaming as Zstd
-import Data.Bits
+import Control.Monad.IO.Class (MonadIO (liftIO))
+import Data.Bits (shiftR, (.&.))
 import qualified Data.ByteString as BS
-import qualified Data.ByteString.Lazy as LB
-import Data.Int
-import Data.Maybe (fromMaybe)
+import Data.Int (Int32, Int64)
+import Data.Maybe (fromJust, fromMaybe)
+import qualified Data.Text as T
+import Data.Text.Encoding (decodeUtf8Lenient)
+import Data.Time (UTCTime)
+import qualified Data.Vector as VB
 import qualified Data.Vector.Unboxed as VU
-import DataFrame.IO.Parquet.Binary
-import DataFrame.IO.Parquet.Thrift
-import DataFrame.IO.Parquet.Types
+import DataFrame.IO.Parquet.Decompress (decompressData)
+import DataFrame.IO.Parquet.Dictionary (
+    DictVals (..),
+    readDictVals,
+ )
+import DataFrame.IO.Parquet.Encoding (decodeDictIndicesV)
+import DataFrame.IO.Parquet.Levels (readLevelsV1V, readLevelsV2V)
+import DataFrame.IO.Parquet.Thrift (
+    ColumnChunk (..),
+    ColumnMetaData (..),
+    CompressionCodec,
+    DataPageHeader (..),
+    DataPageHeaderV2 (..),
+    DictionaryPageHeader (..),
+    Encoding (..),
+    PageHeader (..),
+    PageType (..),
+    ThriftType (..),
+    unField,
+ )
+import DataFrame.IO.Parquet.Time (int96ToUTCTime)
+import DataFrame.IO.Parquet.Utils (ColumnDescription (..))
+import DataFrame.IO.Utils.RandomAccess (RandomAccess (..), Range (Range))
 import DataFrame.Internal.Binary (
     littleEndianInt32,
     littleEndianWord32,
     littleEndianWord64,
  )
-import GHC.Float
-import qualified Snappy
+import GHC.Float (castWord32ToFloat, castWord64ToDouble)
+import Pinch (decodeWithLeftovers)
+import qualified Pinch
+import Streamly.Internal.Data.Unfold (Step (..), Unfold, mkUnfoldM)
 
-isDataPage :: Page -> Bool
-isDataPage page = case pageTypeHeader (pageHeader page) of
-    DataPageHeader{} -> True
-    DataPageHeaderV2{} -> True
-    _ -> False
+-- ---------------------------------------------------------------------------
+-- Types
+-- ---------------------------------------------------------------------------
 
-isDictionaryPage :: Page -> Bool
-isDictionaryPage page = case pageTypeHeader (pageHeader page) of
-    DictionaryPageHeader{} -> True
-    _ -> False
+{- | A type-specific page decoder.
+Given the optional dictionary, the page encoding, the number of present
+values, and the decompressed value bytes, returns exactly @nPresent@ values.
+-}
+type PageDecoder a =
+    Maybe DictVals -> Encoding -> Int -> BS.ByteString -> VB.Vector a
 
-decompressData :: CompressionCodec -> BS.ByteString -> IO BS.ByteString
-decompressData codec compressed = case codec of
-    ZSTD -> do
-        result <- Zstd.decompress
-        drainZstd result compressed []
-      where
-        drainZstd (Zstd.Consume f) input acc = do
-            result <- f input
-            drainZstd result BS.empty acc
-        drainZstd (Zstd.Produce chunk next) _ acc = do
-            result <- next
-            drainZstd result BS.empty (chunk : acc)
-        drainZstd (Zstd.Done final) _ acc =
-            pure $ BS.concat (reverse (final : acc))
-        drainZstd (Zstd.Error msg msg2) _ _ =
-            error ("ZSTD error: " ++ msg ++ " " ++ msg2)
-    SNAPPY -> case Snappy.decompress compressed of
-        Left e -> error (show e)
-        Right res -> pure res
-    UNCOMPRESSED -> pure compressed
-    GZIP -> pure (LB.toStrict (GZip.decompress (BS.fromStrict compressed)))
-    other -> error ("Unsupported compression type: " ++ show other)
+-- ---------------------------------------------------------------------------
+-- Per-type decoders
+-- ---------------------------------------------------------------------------
 
-readPage :: CompressionCodec -> BS.ByteString -> IO (Maybe Page, BS.ByteString)
-readPage c columnBytes =
-    if BS.null columnBytes
-        then pure (Nothing, BS.empty)
-        else do
-            let (hdr, remainder) = readPageHeader emptyPageHeader columnBytes 0
-
-            let compressed = BS.take (fromIntegral $ compressedPageSize hdr) remainder
-
-            fullData <- decompressData c compressed
-
-            pure
-                ( Just $ Page hdr fullData
-                , BS.drop (fromIntegral $ compressedPageSize hdr) remainder
-                )
-
-readPageHeader ::
-    PageHeader -> BS.ByteString -> Int16 -> (PageHeader, BS.ByteString)
-readPageHeader hdr xs lastFieldId =
-    if BS.null xs
-        then (hdr, BS.empty)
-        else
-            let
-                fieldContents = readField' xs lastFieldId
-             in
-                case fieldContents of
-                    Nothing -> (hdr, BS.drop 1 xs)
-                    Just (remainder, _elemType, identifier) -> case identifier of
-                        1 ->
-                            let
-                                (pType, remainder') = readInt32FromBytes remainder
-                             in
-                                readPageHeader
-                                    (hdr{pageHeaderPageType = pageTypeFromInt pType})
-                                    remainder'
-                                    identifier
-                        2 ->
-                            let
-                                (parsedUncompressedPageSize, remainder') = readInt32FromBytes remainder
-                             in
-                                readPageHeader
-                                    (hdr{uncompressedPageSize = parsedUncompressedPageSize})
-                                    remainder'
-                                    identifier
-                        3 ->
-                            let
-                                (parsedCompressedPageSize, remainder') = readInt32FromBytes remainder
-                             in
-                                readPageHeader
-                                    (hdr{compressedPageSize = parsedCompressedPageSize})
-                                    remainder'
-                                    identifier
-                        4 ->
-                            let
-                                (crc, remainder') = readInt32FromBytes remainder
-                             in
-                                readPageHeader (hdr{pageHeaderCrcChecksum = crc}) remainder' identifier
-                        5 ->
-                            let
-                                (dataPageHeader, remainder') = readPageTypeHeader emptyDataPageHeader remainder 0
-                             in
-                                readPageHeader (hdr{pageTypeHeader = dataPageHeader}) remainder' identifier
-                        6 -> error "Index page header not supported"
-                        7 ->
-                            let
-                                (dictionaryPageHeader, remainder') = readPageTypeHeader emptyDictionaryPageHeader remainder 0
-                             in
-                                readPageHeader
-                                    (hdr{pageTypeHeader = dictionaryPageHeader})
-                                    remainder'
-                                    identifier
-                        8 ->
-                            let
-                                (dataPageHeaderV2, remainder') = readPageTypeHeader emptyDataPageHeaderV2 remainder 0
-                             in
-                                readPageHeader (hdr{pageTypeHeader = dataPageHeaderV2}) remainder' identifier
-                        n -> error $ "Unknown page header field " ++ show n
-
-readPageTypeHeader ::
-    PageTypeHeader -> BS.ByteString -> Int16 -> (PageTypeHeader, BS.ByteString)
-readPageTypeHeader INDEX_PAGE_HEADER _ _ = error "readPageTypeHeader: unsupported INDEX_PAGE_HEADER"
-readPageTypeHeader PAGE_TYPE_HEADER_UNKNOWN _ _ = error "readPageTypeHeader: unsupported PAGE_TYPE_HEADER_UNKNOWN"
-readPageTypeHeader hdr@(DictionaryPageHeader{}) xs lastFieldId =
-    if BS.null xs
-        then (hdr, BS.empty)
-        else
-            let
-                fieldContents = readField' xs lastFieldId
-             in
-                case fieldContents of
-                    Nothing -> (hdr, BS.drop 1 xs)
-                    Just (remainder, _elemType, identifier) -> case identifier of
-                        1 ->
-                            let
-                                (numValues, remainder') = readInt32FromBytes remainder
-                             in
-                                readPageTypeHeader
-                                    (hdr{dictionaryPageHeaderNumValues = numValues})
-                                    remainder'
-                                    identifier
-                        2 ->
-                            let
-                                (enc, remainder') = readInt32FromBytes remainder
-                             in
-                                readPageTypeHeader
-                                    (hdr{dictionaryPageHeaderEncoding = parquetEncodingFromInt enc})
-                                    remainder'
-                                    identifier
-                        3 ->
-                            let
-                                isSorted = fromMaybe (error "readPageTypeHeader: not enough bytes") (remainder BS.!? 0)
-                             in
-                                readPageTypeHeader
-                                    (hdr{dictionaryPageIsSorted = isSorted == compactBooleanTrue})
-                                    -- TODO(mchavinda): The bool logic here is a little tricky.
-                                    -- If the field is a bool then you can get the value
-                                    -- from the byte (and you don't have to drop a field).
-                                    -- But in other cases you do.
-                                    -- This might become a problem later but in the mean
-                                    -- time I'm not dropping (this assumes this is the common case).
-                                    remainder
-                                    identifier
-                        n ->
-                            error $ "readPageTypeHeader: unsupported identifier " ++ show n
-readPageTypeHeader hdr@(DataPageHeader{}) xs lastFieldId =
-    if BS.null xs
-        then (hdr, BS.empty)
-        else
-            let
-                fieldContents = readField' xs lastFieldId
-             in
-                case fieldContents of
-                    Nothing -> (hdr, BS.drop 1 xs)
-                    Just (remainder, _elemType, identifier) -> case identifier of
-                        1 ->
-                            let
-                                (numValues, remainder') = readInt32FromBytes remainder
-                             in
-                                readPageTypeHeader
-                                    (hdr{dataPageHeaderNumValues = numValues})
-                                    remainder'
-                                    identifier
-                        2 ->
-                            let
-                                (enc, remainder') = readInt32FromBytes remainder
-                             in
-                                readPageTypeHeader
-                                    (hdr{dataPageHeaderEncoding = parquetEncodingFromInt enc})
-                                    remainder'
-                                    identifier
-                        3 ->
-                            let
-                                (enc, remainder') = readInt32FromBytes remainder
-                             in
-                                readPageTypeHeader
-                                    (hdr{definitionLevelEncoding = parquetEncodingFromInt enc})
-                                    remainder'
-                                    identifier
-                        4 ->
-                            let
-                                (enc, remainder') = readInt32FromBytes remainder
-                             in
-                                readPageTypeHeader
-                                    (hdr{repetitionLevelEncoding = parquetEncodingFromInt enc})
-                                    remainder'
-                                    identifier
-                        5 ->
-                            let
-                                (stats, remainder') = readStatisticsFromBytes emptyColumnStatistics remainder 0
-                             in
-                                readPageTypeHeader (hdr{dataPageHeaderStatistics = stats}) remainder' identifier
-                        n -> error $ show n
-readPageTypeHeader hdr@(DataPageHeaderV2{}) xs lastFieldId =
-    if BS.null xs
-        then (hdr, BS.empty)
-        else
-            let
-                fieldContents = readField' xs lastFieldId
-             in
-                case fieldContents of
-                    Nothing -> (hdr, BS.drop 1 xs)
-                    Just (remainder, _elemType, identifier) -> case identifier of
-                        1 ->
-                            let
-                                (numValues, remainder') = readInt32FromBytes remainder
-                             in
-                                readPageTypeHeader
-                                    (hdr{dataPageHeaderV2NumValues = numValues})
-                                    remainder'
-                                    identifier
-                        2 ->
-                            let
-                                (numNulls, remainder') = readInt32FromBytes remainder
-                             in
-                                readPageTypeHeader
-                                    (hdr{dataPageHeaderV2NumNulls = numNulls})
-                                    remainder'
-                                    identifier
-                        3 ->
-                            let
-                                (parsedNumRows, remainder') = readInt32FromBytes remainder
-                             in
-                                readPageTypeHeader
-                                    (hdr{dataPageHeaderV2NumRows = parsedNumRows})
-                                    remainder'
-                                    identifier
-                        4 ->
-                            let
-                                (enc, remainder') = readInt32FromBytes remainder
-                             in
-                                readPageTypeHeader
-                                    (hdr{dataPageHeaderV2Encoding = parquetEncodingFromInt enc})
-                                    remainder'
-                                    identifier
-                        5 ->
-                            let
-                                (n, remainder') = readInt32FromBytes remainder
-                             in
-                                readPageTypeHeader (hdr{definitionLevelByteLength = n}) remainder' identifier
-                        6 ->
-                            let
-                                (n, remainder') = readInt32FromBytes remainder
-                             in
-                                readPageTypeHeader (hdr{repetitionLevelByteLength = n}) remainder' identifier
-                        7 ->
-                            let
-                                (isCompressed, remainder') = case BS.uncons remainder of
-                                    Just (b, bytes) -> ((b .&. 0x0f) == compactBooleanTrue, bytes)
-                                    Nothing -> (True, BS.empty)
-                             in
-                                readPageTypeHeader
-                                    (hdr{dataPageHeaderV2IsCompressed = isCompressed})
-                                    remainder'
-                                    identifier
-                        8 ->
-                            let
-                                (stats, remainder') = readStatisticsFromBytes emptyColumnStatistics remainder 0
-                             in
-                                readPageTypeHeader
-                                    (hdr{dataPageHeaderV2Statistics = stats})
-                                    remainder'
-                                    identifier
-                        n -> error $ show n
-
-readField' :: BS.ByteString -> Int16 -> Maybe (BS.ByteString, TType, Int16)
-readField' bs lastFieldId = case BS.uncons bs of
-    Nothing -> Nothing
-    Just (x, xs) ->
-        if x .&. 0x0f == 0
-            then Nothing
-            else
-                let modifier = fromIntegral ((x .&. 0xf0) `shiftR` 4) :: Int16
-                    (identifier, remainder) =
-                        if modifier == 0
-                            then readIntFromBytes @Int16 xs
-                            else (lastFieldId + modifier, xs)
-                    elemType = toTType (x .&. 0x0f)
-                 in Just (remainder, elemType, identifier)
-
-readAllPages :: CompressionCodec -> BS.ByteString -> IO [Page]
-readAllPages codec bytes = go bytes []
+boolDecoder :: PageDecoder Bool
+boolDecoder mDict enc nPresent bs = case enc of
+    PLAIN _ -> VB.fromList (readNBool nPresent bs)
+    RLE_DICTIONARY _ -> lookupDict mDict nPresent bs getBool
+    PLAIN_DICTIONARY _ -> lookupDict mDict nPresent bs getBool
+    _ -> error ("boolDecoder: unsupported encoding " ++ show enc)
   where
-    go bs acc =
-        if BS.null bs
-            then return (reverse acc)
-            else do
-                (maybePage, remainderaining) <- readPage codec bs
-                case maybePage of
-                    Nothing -> return (reverse acc)
-                    Just page -> go remainderaining (page : acc)
+    getBool (DBool ds) i = ds VB.! i
+    getBool d _ = error ("boolDecoder: wrong dict type, got " ++ show d)
 
--- | Read n Int32 values directly into an unboxed vector (no intermediate list).
-readNInt32Vec :: Int -> BS.ByteString -> VU.Vector Int32
-readNInt32Vec n bs = VU.generate n (\i -> littleEndianInt32 (BS.drop (4 * i) bs))
+int32Decoder :: PageDecoder Int32
+int32Decoder mDict enc nPresent bs = case enc of
+    PLAIN _ -> VB.convert (readNInt32 nPresent bs)
+    RLE_DICTIONARY _ -> lookupDict mDict nPresent bs getInt32
+    PLAIN_DICTIONARY _ -> lookupDict mDict nPresent bs getInt32
+    _ -> error ("int32Decoder: unsupported encoding " ++ show enc)
+  where
+    getInt32 (DInt32 ds) i = ds VB.! i
+    getInt32 d _ = error ("int32Decoder: wrong dict type, got " ++ show d)
 
--- | Read n Int64 values directly into an unboxed vector.
-readNInt64Vec :: Int -> BS.ByteString -> VU.Vector Int64
-readNInt64Vec n bs = VU.generate n (\i -> fromIntegral (littleEndianWord64 (BS.drop (8 * i) bs)))
+int64Decoder :: PageDecoder Int64
+int64Decoder mDict enc nPresent bs = case enc of
+    PLAIN _ -> VB.convert (readNInt64 nPresent bs)
+    RLE_DICTIONARY _ -> lookupDict mDict nPresent bs getInt64
+    PLAIN_DICTIONARY _ -> lookupDict mDict nPresent bs getInt64
+    _ -> error ("int64Decoder: unsupported encoding " ++ show enc)
+  where
+    getInt64 (DInt64 ds) i = ds VB.! i
+    getInt64 d _ = error ("int64Decoder: wrong dict type, got " ++ show d)
 
--- | Read n Float values directly into an unboxed vector.
-readNFloatVec :: Int -> BS.ByteString -> VU.Vector Float
-readNFloatVec n bs =
-    VU.generate
-        n
-        (\i -> castWord32ToFloat (littleEndianWord32 (BS.drop (4 * i) bs)))
+int96Decoder :: PageDecoder UTCTime
+int96Decoder mDict enc nPresent bs = case enc of
+    PLAIN _ -> VB.fromList (readNInt96 nPresent bs)
+    RLE_DICTIONARY _ -> lookupDict mDict nPresent bs getInt96
+    PLAIN_DICTIONARY _ -> lookupDict mDict nPresent bs getInt96
+    _ -> error ("int96Decoder: unsupported encoding " ++ show enc)
+  where
+    getInt96 (DInt96 ds) i = ds VB.! i
+    getInt96 d _ = error ("int96Decoder: wrong dict type, got " ++ show d)
 
--- | Read n Double values directly into an unboxed vector.
-readNDoubleVec :: Int -> BS.ByteString -> VU.Vector Double
-readNDoubleVec n bs =
-    VU.generate
-        n
-        (\i -> castWord64ToDouble (littleEndianWord64 (BS.drop (8 * i) bs)))
+floatDecoder :: PageDecoder Float
+floatDecoder mDict enc nPresent bs = case enc of
+    PLAIN _ -> VB.convert (readNFloat nPresent bs)
+    RLE_DICTIONARY _ -> lookupDict mDict nPresent bs getFloat
+    PLAIN_DICTIONARY _ -> lookupDict mDict nPresent bs getFloat
+    _ -> error ("floatDecoder: unsupported encoding " ++ show enc)
+  where
+    getFloat (DFloat ds) i = ds VB.! i
+    getFloat d _ = error ("floatDecoder: wrong dict type, got " ++ show d)
 
-readNInt32 :: Int -> BS.ByteString -> ([Int32], BS.ByteString)
-readNInt32 0 bs = ([], bs)
-readNInt32 k bs =
-    let x = littleEndianInt32 (BS.take 4 bs)
-        bs' = BS.drop 4 bs
-        (xs, rest) = readNInt32 (k - 1) bs'
-     in (x : xs, rest)
+doubleDecoder :: PageDecoder Double
+doubleDecoder mDict enc nPresent bs = case enc of
+    PLAIN _ -> VB.convert (readNDouble nPresent bs)
+    RLE_DICTIONARY _ -> lookupDict mDict nPresent bs getDouble
+    PLAIN_DICTIONARY _ -> lookupDict mDict nPresent bs getDouble
+    _ -> error ("doubleDecoder: unsupported encoding " ++ show enc)
+  where
+    getDouble (DDouble ds) i = ds VB.! i
+    getDouble d _ = error ("doubleDecoder: wrong dict type, got " ++ show d)
 
-readNDouble :: Int -> BS.ByteString -> ([Double], BS.ByteString)
-readNDouble 0 bs = ([], bs)
-readNDouble k bs =
-    let x = castWord64ToDouble (littleEndianWord64 (BS.take 8 bs))
-        bs' = BS.drop 8 bs
-        (xs, rest) = readNDouble (k - 1) bs'
-     in (x : xs, rest)
+byteArrayDecoder :: PageDecoder T.Text
+byteArrayDecoder mDict enc nPresent bs = case enc of
+    PLAIN _ -> VB.fromList (readNTexts nPresent bs)
+    RLE_DICTIONARY _ -> lookupDict mDict nPresent bs getText
+    PLAIN_DICTIONARY _ -> lookupDict mDict nPresent bs getText
+    _ -> error ("byteArrayDecoder: unsupported encoding " ++ show enc)
+  where
+    getText (DText ds) i = ds VB.! i
+    getText d _ = error ("byteArrayDecoder: wrong dict type, got " ++ show d)
 
-readNByteArrays :: Int -> BS.ByteString -> ([BS.ByteString], BS.ByteString)
-readNByteArrays 0 bs = ([], bs)
-readNByteArrays k bs =
-    let len = fromIntegral (littleEndianInt32 (BS.take 4 bs)) :: Int
-        body = BS.take len (BS.drop 4 bs)
-        bs' = BS.drop (4 + len) bs
-        (xs, rest) = readNByteArrays (k - 1) bs'
-     in (body : xs, rest)
+fixedLenByteArrayDecoder :: Int -> PageDecoder T.Text
+fixedLenByteArrayDecoder len mDict enc nPresent bs = case enc of
+    PLAIN _ -> VB.fromList (readNFixedTexts len nPresent bs)
+    RLE_DICTIONARY _ -> lookupDict mDict nPresent bs getText
+    PLAIN_DICTIONARY _ -> lookupDict mDict nPresent bs getText
+    _ -> error ("fixedLenByteArrayDecoder: unsupported encoding " ++ show enc)
+  where
+    getText (DText ds) i = ds VB.! i
+    getText d _ = error ("fixedLenByteArrayDecoder: wrong dict type, got " ++ show d)
 
-readNBool :: Int -> BS.ByteString -> ([Bool], BS.ByteString)
-readNBool 0 bs = ([], bs)
+{- | Shared dictionary-path helper: decode @nPresent@ RLE/bit-packed indices
+and look each one up in the dictionary.
+-}
+lookupDict ::
+    Maybe DictVals ->
+    Int ->
+    BS.ByteString ->
+    (DictVals -> Int -> a) ->
+    VB.Vector a
+lookupDict mDict nPresent bs f = case mDict of
+    Nothing -> error "Dictionary-encoded page but no dictionary page seen"
+    Just dict ->
+        let (idxs, _) = decodeDictIndicesV nPresent bs
+         in VB.generate nPresent (f dict . VU.unsafeIndex idxs)
+
+-- ---------------------------------------------------------------------------
+-- Core page-iteration loop
+-- ---------------------------------------------------------------------------
+
+-- | Read the raw (compressed) byte range for a column chunk.
+readChunkBytes ::
+    (RandomAccess m) =>
+    ColumnChunk ->
+    m (CompressionCodec, ThriftType, BS.ByteString)
+readChunkBytes columnChunk = do
+    let meta = fromJust . unField $ columnChunk.cc_meta_data
+        codec = unField meta.cmd_codec
+        pType = unField meta.cmd_type
+        dataOffset = fromIntegral . unField $ meta.cmd_data_page_offset
+        dictOffset = fromIntegral <$> unField meta.cmd_dictionary_page_offset
+        offset = fromMaybe dataOffset dictOffset
+        compLen = fromIntegral . unField $ meta.cmd_total_compressed_size
+    rawBytes <- readBytes (Range offset compLen)
+    return (codec, pType, rawBytes)
+
+{- | An 'Unfold' from a 'ColumnChunk' to per-page value triples.
+
+The seed is a 'ColumnChunk'.  The inject step reads the chunk's compressed
+bytes and discovers the codec and physical type from the column metadata.
+Codec and type are then threaded through the unfold state along with the
+running dictionary and remaining bytes, so no intermediate list or
+concatenation step is needed.  Use with 'Stream.unfoldEach' to produce a
+flat stream of per-page results directly from a stream of column chunks.
+
+Dictionary pages are consumed silently and update the running dictionary
+that is threaded through the unfold state.
+
+The internal state is
+@(Maybe DictVals, BS.ByteString, CompressionCodec, ThriftType)@.
+
+-- TODO: when a page index is available, use it here to compute which page
+-- byte ranges to request from the RandomAccess layer instead of reading the
+-- entire column chunk in one contiguous read.
+
+-- TODO: accept an optional row-range and use the column/offset page index
+-- (when present in file metadata) to Skip pages whose row range does not
+-- overlap the requested range, avoiding decompression of irrelevant pages
+-- entirely.
+-}
+readPages ::
+    (RandomAccess m, MonadIO m) =>
+    ColumnDescription ->
+    PageDecoder a ->
+    Unfold m ColumnChunk (VB.Vector a, VU.Vector Int, VU.Vector Int)
+readPages description decoder = mkUnfoldM step inject
+  where
+    maxDef = fromIntegral description.maxDefinitionLevel :: Int
+    maxRep = fromIntegral description.maxRepetitionLevel :: Int
+
+    -- Inject: read chunk bytes; put codec and pType into state.
+    inject cc = do
+        (codec, pType, rawBytes) <- readChunkBytes cc
+        return (Nothing, rawBytes, codec, pType)
+
+    step (dict, bs, codec, pType)
+        | BS.null bs = return Stop
+        | otherwise = case parsePageHeader bs of
+            Left e -> error ("readPages: failed to parse page header: " ++ e)
+            Right (rest, hdr) -> do
+                let compSz = fromIntegral . unField $ hdr.ph_compressed_page_size
+                    uncmpSz = fromIntegral . unField $ hdr.ph_uncompressed_page_size
+                    (pageData, rest') = BS.splitAt compSz rest
+                case unField hdr.ph_type of
+                    DICTIONARY_PAGE _ -> do
+                        let dictHdr =
+                                fromMaybe
+                                    (error "DICTIONARY_PAGE: missing dictionary page header")
+                                    (unField hdr.ph_dictionary_page_header)
+                            numVals = unField dictHdr.diph_num_values
+                        decompressed <- liftIO $ decompressData uncmpSz codec pageData
+                        let d = readDictVals pType decompressed numVals description.typeLength
+                        return $ Skip (Just d, rest', codec, pType)
+                    DATA_PAGE _ -> do
+                        let dph =
+                                fromMaybe
+                                    (error "DATA_PAGE: missing data page header")
+                                    (unField hdr.ph_data_page_header)
+                            n = fromIntegral . unField $ dph.dph_num_values
+                            enc = unField dph.dph_encoding
+                        decompressed <- liftIO $ decompressData uncmpSz codec pageData
+                        let (defLvls, repLvls, nPresent, valBytes) =
+                                readLevelsV1V n maxDef maxRep decompressed
+                            triple = (decoder dict enc nPresent valBytes, defLvls, repLvls)
+                        return $ Yield triple (dict, rest', codec, pType)
+                    DATA_PAGE_V2 _ -> do
+                        let dph2 =
+                                fromMaybe
+                                    (error "DATA_PAGE_V2: missing data page header v2")
+                                    (unField hdr.ph_data_page_header_v2)
+                            n = fromIntegral . unField $ dph2.dph2_num_values
+                            enc = unField dph2.dph2_encoding
+                            defLen = unField dph2.dph2_definition_levels_byte_length
+                            repLen = unField dph2.dph2_repetition_levels_byte_length
+                            -- V2: levels are never compressed; only the value
+                            -- payload is (optionally) compressed.
+                            isCompressed = fromMaybe True (unField dph2.dph2_is_compressed)
+                            (defLvls, repLvls, nPresent, compValBytes) =
+                                readLevelsV2V n maxDef maxRep repLen defLen pageData
+                        valBytes <-
+                            if isCompressed
+                                then liftIO $ decompressData uncmpSz codec compValBytes
+                                else pure compValBytes
+                        let triple = (decoder dict enc nPresent valBytes, defLvls, repLvls)
+                        return $ Yield triple (dict, rest', codec, pType)
+                    INDEX_PAGE _ -> return $ Skip (dict, rest', codec, pType)
+
+-- ---------------------------------------------------------------------------
+-- Page header parsing
+-- ---------------------------------------------------------------------------
+
+parsePageHeader :: BS.ByteString -> Either String (BS.ByteString, PageHeader)
+parsePageHeader = decodeWithLeftovers Pinch.compactProtocol
+
+-- ---------------------------------------------------------------------------
+-- Batch value readers
+-- ---------------------------------------------------------------------------
+
+readNBool :: Int -> BS.ByteString -> [Bool]
 readNBool count bs =
     let totalBytes = (count + 7) `div` 8
-        chunk = BS.take totalBytes bs
-        rest = BS.drop totalBytes bs
         bits =
             concatMap
                 (\b -> map (\i -> (b `shiftR` i) .&. 1 == 1) [0 .. 7])
-                (BS.unpack chunk)
-        bools = take count bits
-     in (bools, rest)
+                (BS.unpack (BS.take totalBytes bs))
+     in take count bits
 
-readNInt64 :: Int -> BS.ByteString -> ([Int64], BS.ByteString)
-readNInt64 0 bs = ([], bs)
-readNInt64 k bs =
-    let x = fromIntegral (littleEndianWord64 (BS.take 8 bs))
-        bs' = BS.drop 8 bs
-        (xs, rest) = readNInt64 (k - 1) bs'
-     in (x : xs, rest)
+readNInt32 :: Int -> BS.ByteString -> VU.Vector Int32
+readNInt32 n bs = VU.generate n $ \i -> littleEndianInt32 (BS.drop (4 * i) bs)
 
-readNFloat :: Int -> BS.ByteString -> ([Float], BS.ByteString)
-readNFloat 0 bs = ([], bs)
-readNFloat k bs =
-    let x = castWord32ToFloat (littleEndianWord32 (BS.take 4 bs))
-        bs' = BS.drop 4 bs
-        (xs, rest) = readNFloat (k - 1) bs'
-     in (x : xs, rest)
+readNInt64 :: Int -> BS.ByteString -> VU.Vector Int64
+readNInt64 n bs = VU.generate n $ \i ->
+    fromIntegral (littleEndianWord64 (BS.drop (8 * i) bs))
 
-splitFixed :: Int -> Int -> BS.ByteString -> ([BS.ByteString], BS.ByteString)
-splitFixed 0 _ bs = ([], bs)
-splitFixed k len bs =
-    let body = BS.take len bs
-        bs' = BS.drop len bs
-        (xs, rest) = splitFixed (k - 1) len bs'
-     in (body : xs, rest)
+readNInt96 :: Int -> BS.ByteString -> [UTCTime]
+readNInt96 0 _ = []
+readNInt96 n bs = int96ToUTCTime (BS.take 12 bs) : readNInt96 (n - 1) (BS.drop 12 bs)
 
-readStatisticsFromBytes ::
-    ColumnStatistics -> BS.ByteString -> Int16 -> (ColumnStatistics, BS.ByteString)
-readStatisticsFromBytes cs xs lastFieldId =
-    let
-        fieldContents = readField' xs lastFieldId
-     in
-        case fieldContents of
-            Nothing -> (cs, BS.drop 1 xs)
-            Just (remainder, _elemType, identifier) -> case identifier of
-                1 ->
-                    let
-                        (maxInBytes, remainder') = readByteStringFromBytes remainder
-                     in
-                        readStatisticsFromBytes (cs{columnMax = maxInBytes}) remainder' identifier
-                2 ->
-                    let
-                        (minInBytes, remainder') = readByteStringFromBytes remainder
-                     in
-                        readStatisticsFromBytes (cs{columnMin = minInBytes}) remainder' identifier
-                3 ->
-                    let
-                        (nullCount, remainder') = readIntFromBytes @Int64 remainder
-                     in
-                        readStatisticsFromBytes (cs{columnNullCount = nullCount}) remainder' identifier
-                4 ->
-                    let
-                        (distinctCount, remainder') = readIntFromBytes @Int64 remainder
-                     in
-                        readStatisticsFromBytes
-                            (cs{columnDistictCount = distinctCount})
-                            remainder'
-                            identifier
-                5 ->
-                    let
-                        (maxInBytes, remainder') = readByteStringFromBytes remainder
-                     in
-                        readStatisticsFromBytes (cs{columnMaxValue = maxInBytes}) remainder' identifier
-                6 ->
-                    let
-                        (minInBytes, remainder') = readByteStringFromBytes remainder
-                     in
-                        readStatisticsFromBytes (cs{columnMinValue = minInBytes}) remainder' identifier
-                7 ->
-                    case BS.uncons remainder of
-                        Nothing ->
-                            error "readStatisticsFromBytes: not enough bytes"
-                        Just (isMaxValueExact, remainder') ->
-                            readStatisticsFromBytes
-                                (cs{isColumnMaxValueExact = isMaxValueExact == compactBooleanTrue})
-                                remainder'
-                                identifier
-                8 ->
-                    case BS.uncons remainder of
-                        Nothing ->
-                            error "readStatisticsFromBytes: not enough bytes"
-                        Just (isMinValueExact, remainder') ->
-                            readStatisticsFromBytes
-                                (cs{isColumnMinValueExact = isMinValueExact == compactBooleanTrue})
-                                remainder'
-                                identifier
-                n -> error $ show n
+readNFloat :: Int -> BS.ByteString -> VU.Vector Float
+readNFloat n bs = VU.generate n $ \i ->
+    castWord32ToFloat (littleEndianWord32 (BS.drop (4 * i) bs))
+
+readNDouble :: Int -> BS.ByteString -> VU.Vector Double
+readNDouble n bs = VU.generate n $ \i ->
+    castWord64ToDouble (littleEndianWord64 (BS.drop (8 * i) bs))
+
+readNTexts :: Int -> BS.ByteString -> [T.Text]
+readNTexts 0 _ = []
+readNTexts n bs =
+    let len = fromIntegral . littleEndianInt32 . BS.take 4 $ bs
+        text = decodeUtf8Lenient . BS.take len . BS.drop 4 $ bs
+     in text : readNTexts (n - 1) (BS.drop (4 + len) bs)
+
+readNFixedTexts :: Int -> Int -> BS.ByteString -> [T.Text]
+readNFixedTexts _ 0 _ = []
+readNFixedTexts len n bs =
+    decodeUtf8Lenient (BS.take len bs)
+        : readNFixedTexts len (n - 1) (BS.drop len bs)
