+++ src/DataFrame/IO/Parquet/Thrift.hs
@@ -1,1199 +1,584 @@
-{-# LANGUAGE AllowAmbiguousTypes #-}
-{-# LANGUAGE GADTs #-}
-{-# LANGUAGE OverloadedStrings #-}
-{-# LANGUAGE ScopedTypeVariables #-}
-{-# LANGUAGE StrictData #-}
-{-# LANGUAGE TypeApplications #-}
+{-# LANGUAGE DataKinds #-}
+{-# LANGUAGE DeriveGeneric #-}
+{-# LANGUAGE TypeFamilies #-}
 
 module DataFrame.IO.Parquet.Thrift where
 
-import Control.Monad
-import Data.Bits
-import qualified Data.ByteString as BS
-import Data.Char
-import Data.IORef
-import Data.Int
-import qualified Data.Map as M
-import Data.Maybe
-import qualified Data.Text as T
-import Data.Typeable (Typeable)
-import qualified Data.Vector as V
-import qualified Data.Vector.Unboxed as VU
-import Data.Word
-import DataFrame.IO.Parquet.Binary
-import DataFrame.IO.Parquet.Seeking
-import DataFrame.IO.Parquet.Types
-import qualified DataFrame.Internal.Column as DI
-import DataFrame.Internal.DataFrame (DataFrame, unsafeGetColumn)
-import qualified DataFrame.Operations.Core as DI
-import Type.Reflection (
-    eqTypeRep,
-    typeRep,
-    (:~~:) (HRefl),
- )
+import Data.ByteString (ByteString)
+import Data.Int (Int16, Int32, Int64, Int8)
+import Data.Text (Text)
+import GHC.Generics (Generic)
+import GHC.TypeLits (KnownNat)
+import Pinch (Enumeration, Field, Pinchable (..))
+import qualified Pinch
 
-data SchemaElement = SchemaElement
-    { elementName :: T.Text
-    , elementType :: TType
-    , typeLength :: Int32
-    , numChildren :: Int32
-    , fieldId :: Int32
-    , repetitionType :: RepetitionType
-    , convertedType :: Int32
-    , scale :: Int32
-    , precision :: Int32
-    , logicalType :: LogicalType
+-- Primitive Parquet Types
+-- https://github.com/apache/parquet-format/blob/master/src/main/thrift/parquet.thrift#L32
+data ThriftType
+    = BOOLEAN (Enumeration 0)
+    | INT32 (Enumeration 1)
+    | INT64 (Enumeration 2)
+    | INT96 (Enumeration 3)
+    | FLOAT (Enumeration 4)
+    | DOUBLE (Enumeration 5)
+    | BYTE_ARRAY (Enumeration 6)
+    | FIXED_LEN_BYTE_ARRAY (Enumeration 7)
+    deriving (Eq, Show, Generic)
+
+instance Pinchable ThriftType
+
+-- https://github.com/apache/parquet-format/blob/master/src/main/thrift/parquet.thrift#L183
+data FieldRepetitionType
+    = REQUIRED (Enumeration 0)
+    | OPTIONAL (Enumeration 1)
+    | REPEATED (Enumeration 2)
+    deriving (Eq, Show, Generic)
+
+instance Pinchable FieldRepetitionType
+
+-- https://github.com/apache/parquet-format/blob/master/src/main/thrift/parquet.thrift#L203
+data Encoding
+    = PLAIN (Enumeration 0)
+    | -- GROUP_VAR_INT Encoding was never used
+      -- https://github.com/apache/parquet-format/blob/master/src/main/thrift/parquet.thrift#L578
+      PLAIN_DICTIONARY (Enumeration 2)
+    | RLE (Enumeration 3)
+    | BIT_PACKED (Enumeration 4)
+    | DELTA_BINARY_PACKED (Enumeration 5)
+    | DELTA_LENGTH_BYTE_ARRAY (Enumeration 6)
+    | DELTA_BYTE_ARRAY (Enumeration 7)
+    | RLE_DICTIONARY (Enumeration 8)
+    | BYTE_STREAM_SPLIT (Enumeration 9)
+    deriving (Eq, Show, Generic)
+
+instance Pinchable Encoding
+
+-- https://github.com/apache/parquet-format/blob/master/src/main/thrift/parquet.thrift#L244
+data CompressionCodec
+    = UNCOMPRESSED (Enumeration 0)
+    | SNAPPY (Enumeration 1)
+    | GZIP (Enumeration 2)
+    | LZO (Enumeration 3)
+    | BROTLI (Enumeration 4)
+    | LZ4 (Enumeration 5)
+    | ZSTD (Enumeration 6)
+    | LZ4_RAW (Enumeration 7)
+    deriving (Eq, Show, Generic)
+
+instance Pinchable CompressionCodec
+
+-- https://github.com/apache/parquet-format/blob/master/src/main/thrift/parquet.thrift#L261
+data PageType
+    = DATA_PAGE (Enumeration 0)
+    | INDEX_PAGE (Enumeration 1)
+    | DICTIONARY_PAGE (Enumeration 2)
+    | DATA_PAGE_V2 (Enumeration 3)
+    deriving (Eq, Show, Generic)
+
+instance Pinchable PageType
+
+-- https://github.com/apache/parquet-format/blob/master/src/main/thrift/parquet.thrift#L271
+data BoundaryOrder
+    = UNORDERED (Enumeration 0)
+    | ASCENDING (Enumeration 1)
+    | DESCENDING (Enumeration 2)
+    deriving (Eq, Show, Generic)
+
+instance Pinchable BoundaryOrder
+
+-- Logical type annotations
+-- Empty structs can't use deriving Generic with Pinch, so we use a unit-like workaround.
+-- We represent empty structs as a newtype over () with a manual Pinchable instance.
+
+-- https://github.com/apache/parquet-format/blob/master/src/main/thrift/parquet.thrift#L283
+-- struct StringType {}
+data StringType = StringType deriving (Eq, Show)
+instance Pinchable StringType where
+    type Tag StringType = Pinch.TStruct
+    pinch _ = Pinch.struct []
+    unpinch _ = pure StringType
+
+data UUIDType = UUIDType deriving (Eq, Show)
+instance Pinchable UUIDType where
+    type Tag UUIDType = Pinch.TStruct
+    pinch _ = Pinch.struct []
+    unpinch _ = pure UUIDType
+
+data MapType = MapType deriving (Eq, Show)
+instance Pinchable MapType where
+    type Tag MapType = Pinch.TStruct
+    pinch _ = Pinch.struct []
+    unpinch _ = pure MapType
+
+data ListType = ListType deriving (Eq, Show)
+instance Pinchable ListType where
+    type Tag ListType = Pinch.TStruct
+    pinch _ = Pinch.struct []
+    unpinch _ = pure ListType
+
+data EnumType = EnumType deriving (Eq, Show)
+instance Pinchable EnumType where
+    type Tag EnumType = Pinch.TStruct
+    pinch _ = Pinch.struct []
+    unpinch _ = pure EnumType
+
+data DateType = DateType deriving (Eq, Show)
+instance Pinchable DateType where
+    type Tag DateType = Pinch.TStruct
+    pinch _ = Pinch.struct []
+    unpinch _ = pure DateType
+
+data Float16Type = Float16Type deriving (Eq, Show)
+instance Pinchable Float16Type where
+    type Tag Float16Type = Pinch.TStruct
+    pinch _ = Pinch.struct []
+    unpinch _ = pure Float16Type
+
+data NullType = NullType deriving (Eq, Show)
+instance Pinchable NullType where
+    type Tag NullType = Pinch.TStruct
+    pinch _ = Pinch.struct []
+    unpinch _ = pure NullType
+
+data JsonType = JsonType deriving (Eq, Show)
+instance Pinchable JsonType where
+    type Tag JsonType = Pinch.TStruct
+    pinch _ = Pinch.struct []
+    unpinch _ = pure JsonType
+
+data BsonType = BsonType deriving (Eq, Show)
+instance Pinchable BsonType where
+    type Tag BsonType = Pinch.TStruct
+    pinch _ = Pinch.struct []
+    unpinch _ = pure BsonType
+
+data VariantType = VariantType deriving (Eq, Show)
+instance Pinchable VariantType where
+    type Tag VariantType = Pinch.TStruct
+    pinch _ = Pinch.struct []
+    unpinch _ = pure VariantType
+
+-- https://github.com/apache/parquet-format/blob/master/src/main/thrift/parquet.thrift#L290
+data TimeUnit
+    = MILLIS (Field 1 MilliSeconds)
+    | MICROS (Field 2 MicroSeconds)
+    | NANOS (Field 3 NanoSeconds)
+    deriving (Eq, Show, Generic)
+
+instance Pinchable TimeUnit
+
+data MilliSeconds = MilliSeconds deriving (Eq, Show)
+instance Pinchable MilliSeconds where
+    type Tag MilliSeconds = Pinch.TStruct
+    pinch _ = Pinch.struct []
+    unpinch _ = pure MilliSeconds
+
+data MicroSeconds = MicroSeconds deriving (Eq, Show)
+instance Pinchable MicroSeconds where
+    type Tag MicroSeconds = Pinch.TStruct
+    pinch _ = Pinch.struct []
+    unpinch _ = pure MicroSeconds
+
+data NanoSeconds = NanoSeconds deriving (Eq, Show)
+instance Pinchable NanoSeconds where
+    type Tag NanoSeconds = Pinch.TStruct
+    pinch _ = Pinch.struct []
+    unpinch _ = pure NanoSeconds
+
+-- https://github.com/apache/parquet-format/blob/master/src/main/thrift/parquet.thrift#L317
+data DecimalType
+    = DecimalType
+    { decimal_scale :: Field 1 Int32
+    , decimal_precision :: Field 2 Int32
     }
-    deriving (Show, Eq)
+    deriving (Eq, Show, Generic)
 
-createParquetSchema :: DataFrame -> [SchemaElement]
-createParquetSchema df = schemaDef : map toSchemaElement (DI.columnNames df)
-  where
-    -- The schema always contains an initial element
-    -- indicating the group of fields.
-    schemaDef =
-        SchemaElement
-            { elementName = "schema"
-            , elementType = STOP
-            , typeLength = 0
-            , numChildren = fromIntegral (snd (DI.dimensions df))
-            , fieldId = -1
-            , repetitionType = UNKNOWN_REPETITION_TYPE
-            , convertedType = 0
-            , scale = 0
-            , precision = 0
-            , logicalType = LOGICAL_TYPE_UNKNOWN
-            }
-    toSchemaElement colName =
-        let
-            colType :: TType
-            colType = case unsafeGetColumn colName df of
-                (DI.BoxedColumn _ (_col :: V.Vector a)) -> haskellToTType @a
-                (DI.UnboxedColumn _ (_col :: VU.Vector a)) -> haskellToTType @a
-            lType =
-                if DI.hasElemType @T.Text (unsafeGetColumn colName df)
-                    || DI.hasElemType @(Maybe T.Text) (unsafeGetColumn colName df)
-                    then STRING_TYPE
-                    else LOGICAL_TYPE_UNKNOWN
-         in
-            SchemaElement colName colType 0 0 (-1) OPTIONAL 0 0 0 lType
+instance Pinchable DecimalType
 
-data KeyValue = KeyValue
-    { key :: String
-    , value :: String
+-- https://github.com/apache/parquet-format/blob/master/src/main/thrift/parquet.thrift#L328
+data IntType
+    = IntType
+    { int_bitWidth :: Field 1 Int8
+    , int_isSigned :: Field 2 Bool
     }
-    deriving (Show, Eq)
+    deriving (Eq, Show, Generic)
 
-data FileMetadata = FileMetaData
-    { version :: Int32
-    , schema :: [SchemaElement]
-    , numRows :: Integer
-    , rowGroups :: [RowGroup]
-    , keyValueMetadata :: [KeyValue]
-    , createdBy :: Maybe String
-    , columnOrders :: [ColumnOrder]
-    , encryptionAlgorithm :: EncryptionAlgorithm
-    , footerSigningKeyMetadata :: BS.ByteString
+instance Pinchable IntType
+
+-- https://github.com/apache/parquet-format/blob/master/src/main/thrift/parquet.thrift#L338
+data TimeType
+    = TimeType
+    { time_isAdjustedToUTC :: Field 1 Bool
+    , time_unit :: Field 2 TimeUnit
     }
-    deriving (Show, Eq)
+    deriving (Eq, Show, Generic)
 
-data TType
-    = STOP
-    | BOOL
-    | BYTE
-    | I16
-    | I32
-    | I64
-    | I96
-    | FLOAT
-    | DOUBLE
-    | STRING
-    | LIST
-    | SET
-    | MAP
-    | STRUCT
-    | UUID
-    deriving (Show, Eq)
+instance Pinchable TimeType
 
-haskellToTType :: forall a. (Typeable a) => TType
-haskellToTType
-    | is @Bool = BOOL
-    | is @Int8 = BYTE
-    | is @Word8 = BYTE
-    | is @Int16 = I16
-    | is @Word16 = I16
-    | is @Int32 = I32
-    | is @Word32 = I32
-    | is @Int64 = I64
-    | is @Word64 = I64
-    | is @Float = FLOAT
-    | is @Double = DOUBLE
-    | is @String = STRING
-    | is @T.Text = STRING
-    | is @BS.ByteString = STRING
-    | otherwise = STOP
-  where
-    is :: forall x. (Typeable x) => Bool
-    is = case eqTypeRep (typeRep @a) (typeRep @x) of
-        Just HRefl -> True
-        Nothing -> False
-
-defaultMetadata :: FileMetadata
-defaultMetadata =
-    FileMetaData
-        { version = 0
-        , schema = []
-        , numRows = 0
-        , rowGroups = []
-        , keyValueMetadata = []
-        , createdBy = Nothing
-        , columnOrders = []
-        , encryptionAlgorithm = ENCRYPTION_ALGORITHM_UNKNOWN
-        , footerSigningKeyMetadata = BS.empty
-        }
-
-data ColumnMetaData = ColumnMetaData
-    { columnType :: ParquetType
-    , columnEncodings :: [ParquetEncoding]
-    , columnPathInSchema :: [String]
-    , columnCodec :: CompressionCodec
-    , columnNumValues :: Int64
-    , columnTotalUncompressedSize :: Int64
-    , columnTotalCompressedSize :: Int64
-    , columnKeyValueMetadata :: [KeyValue]
-    , columnDataPageOffset :: Int64
-    , columnIndexPageOffset :: Int64
-    , columnDictionaryPageOffset :: Int64
-    , columnStatistics :: ColumnStatistics
-    , columnEncodingStats :: [PageEncodingStats]
-    , bloomFilterOffset :: Int64
-    , bloomFilterLength :: Int32
-    , columnSizeStatistics :: SizeStatistics
-    , columnGeospatialStatistics :: GeospatialStatistics
+-- https://github.com/apache/parquet-format/blob/master/src/main/thrift/parquet.thrift#L349
+data TimestampType
+    = TimestampType
+    { timestamp_isAdjustedToUTC :: Field 1 Bool
+    , timestamp_unit :: Field 2 TimeUnit
     }
-    deriving (Show, Eq)
+    deriving (Eq, Show, Generic)
 
-data ColumnChunk = ColumnChunk
-    { columnChunkFilePath :: String
-    , columnChunkMetadataFileOffset :: Int64
-    , columnMetaData :: ColumnMetaData
-    , columnChunkOffsetIndexOffset :: Int64
-    , columnChunkOffsetIndexLength :: Int32
-    , columnChunkColumnIndexOffset :: Int64
-    , columnChunkColumnIndexLength :: Int32
-    , cryptoMetadata :: ColumnCryptoMetadata
-    , encryptedColumnMetadata :: BS.ByteString
+instance Pinchable TimestampType
+
+-- https://github.com/apache/parquet-format/blob/master/src/main/thrift/parquet.thrift#L360
+-- union LogicalType
+data LogicalType
+    = LT_STRING (Field 1 StringType)
+    | LT_MAP (Field 2 MapType)
+    | LT_LIST (Field 3 ListType)
+    | LT_ENUM (Field 4 EnumType)
+    | LT_DECIMAL (Field 5 DecimalType)
+    | LT_DATE (Field 6 DateType)
+    | LT_TIME (Field 7 TimeType)
+    | LT_TIMESTAMP (Field 8 TimestampType)
+    | LT_INTEGER (Field 10 IntType)
+    | LT_NULL (Field 11 NullType)
+    | LT_JSON (Field 12 JsonType)
+    | LT_BSON (Field 13 BsonType)
+    | LT_UUID (Field 14 UUIDType)
+    | LT_FLOAT16 (Field 15 Float16Type)
+    | LT_VARIANT (Field 16 VariantType)
+    deriving (Eq, Show, Generic)
+
+instance Pinchable LogicalType
+
+-- https://github.com/apache/parquet-format/blob/master/src/main/thrift/parquet.thrift#L270
+data ConvertedType
+    = UTF8 (Enumeration 0)
+    | MAP (Enumeration 1)
+    | MAP_KEY_VALUE (Enumeration 2)
+    | LIST (Enumeration 3)
+    | ENUM (Enumeration 4)
+    | DECIMAL (Enumeration 5)
+    | DATE (Enumeration 6)
+    | TIME_MILLIS (Enumeration 7)
+    | TIME_MICROS (Enumeration 8)
+    | TIMESTAMP_MILLIS (Enumeration 9)
+    | TIMESTAMP_MICROS (Enumeration 10)
+    | UINT_8 (Enumeration 11)
+    | UINT_16 (Enumeration 12)
+    | UINT_32 (Enumeration 13)
+    | UINT_64 (Enumeration 14)
+    | INT_8 (Enumeration 15)
+    | INT_16 (Enumeration 16)
+    | INT_32 (Enumeration 17)
+    | INT_64 (Enumeration 18)
+    | JSON (Enumeration 19)
+    | BSON (Enumeration 20)
+    | INTERVAL (Enumeration 21)
+    deriving (Eq, Show, Generic)
+
+instance Pinchable ConvertedType
+
+-- https://github.com/apache/parquet-format/blob/master/src/main/thrift/parquet.thrift#L505
+data SchemaElement
+    = SchemaElement
+    { schematype :: Field 1 (Maybe ThriftType) -- called just type in parquet.thrift
+    , type_length :: Field 2 (Maybe Int32)
+    , repetition_type :: Field 3 (Maybe FieldRepetitionType)
+    , name :: Field 4 Text
+    , num_children :: Field 5 (Maybe Int32)
+    , converted_type :: Field 6 (Maybe ConvertedType)
+    , scale :: Field 7 (Maybe Int32)
+    , precision :: Field 8 (Maybe Int32)
+    , field_id :: Field 9 (Maybe Int32)
+    , logicalType :: Field 10 (Maybe LogicalType)
     }
-    deriving (Show, Eq)
+    deriving (Eq, Show, Generic)
 
-data RowGroup = RowGroup
-    { rowGroupColumns :: [ColumnChunk]
-    , totalByteSize :: Int64
-    , rowGroupNumRows :: Int64
-    , rowGroupSortingColumns :: [SortingColumn]
-    , fileOffset :: Int64
-    , totalCompressedSize :: Int64
-    , ordinal :: Int16
+instance Pinchable SchemaElement
+
+-- https://github.com/apache/parquet-format/blob/master/src/main/thrift/parquet.thrift#L560
+data Statistics
+    = Statistics
+    { stats_max :: Field 1 (Maybe ByteString)
+    , stats_min :: Field 2 (Maybe ByteString)
+    , stats_null_count :: Field 3 (Maybe Int64)
+    , stats_distinct_count :: Field 4 (Maybe Int64)
+    , stats_max_value :: Field 5 (Maybe ByteString)
+    , stats_min_value :: Field 6 (Maybe ByteString)
+    , stats_is_max_value_exact :: Field 7 (Maybe Bool)
+    , stats_is_min_value_exact :: Field 8 (Maybe Bool)
     }
-    deriving (Show, Eq)
+    deriving (Eq, Show, Generic)
 
-defaultSchemaElement :: SchemaElement
-defaultSchemaElement =
-    SchemaElement
-        ""
-        STOP
-        0
-        0
-        (-1)
-        UNKNOWN_REPETITION_TYPE
-        0
-        0
-        0
-        LOGICAL_TYPE_UNKNOWN
+instance Pinchable Statistics
 
-emptyColumnMetadata :: ColumnMetaData
-emptyColumnMetadata =
-    ColumnMetaData
-        PARQUET_TYPE_UNKNOWN
-        []
-        []
-        COMPRESSION_CODEC_UNKNOWN
-        0
-        0
-        0
-        []
-        0
-        0
-        0
-        emptyColumnStatistics
-        []
-        0
-        0
-        emptySizeStatistics
-        emptyGeospatialStatistics
+-- https://github.com/apache/parquet-format/blob/master/src/main/thrift/parquet.thrift#L600
+data PageEncodingStats
+    = PageEncodingStats
+    { pes_page_type :: Field 1 PageType
+    , pes_encoding :: Field 2 Encoding
+    , pes_count :: Field 3 Int32
+    }
+    deriving (Eq, Show, Generic)
 
-emptyColumnChunk :: ColumnChunk
-emptyColumnChunk =
-    ColumnChunk
-        ""
-        0
-        emptyColumnMetadata
-        0
-        0
-        0
-        0
-        COLUMN_CRYPTO_METADATA_UNKNOWN
-        BS.empty
+instance Pinchable PageEncodingStats
 
-emptyKeyValue :: KeyValue
-emptyKeyValue = KeyValue{key = "", value = ""}
+-- https://github.com/apache/parquet-format/blob/master/src/main/thrift/parquet.thrift#L614
+data ColumnMetaData
+    = ColumnMetaData
+    { cmd_type :: Field 1 ThriftType
+    , cmd_encodings :: Field 2 [Encoding]
+    , cmd_path_in_schema :: Field 3 [Text]
+    , cmd_codec :: Field 4 CompressionCodec
+    , cmd_num_values :: Field 5 Int64
+    , cmd_total_uncompressed_size :: Field 6 Int64
+    , cmd_total_compressed_size :: Field 7 Int64
+    , cmd_key_value_metadata :: Field 8 (Maybe [KeyValue])
+    , cmd_data_page_offset :: Field 9 Int64
+    , cmd_index_page_offset :: Field 10 (Maybe Int64)
+    , cmd_dictionary_page_offset :: Field 11 (Maybe Int64)
+    , cmd_statistics :: Field 12 (Maybe Statistics)
+    , cmd_encoding_stats :: Field 13 (Maybe [PageEncodingStats])
+    , cmd_bloom_filter_offset :: Field 14 (Maybe Int64)
+    , cmd_bloom_filter_length :: Field 15 (Maybe Int32)
+    }
+    deriving (Eq, Show, Generic)
 
-emptyRowGroup :: RowGroup
-emptyRowGroup = RowGroup [] 0 0 [] 0 0 0
+instance Pinchable ColumnMetaData
 
-compactBooleanTrue
-    , compactI32
-    , compactI64
-    , compactDouble
-    , compactBinary
-    , compactList
-    , compactStruct ::
-        Word8
-compactBooleanTrue = 0x01
-compactI32 = 0x05
-compactI64 = 0x06
-compactDouble = 0x07
-compactBinary = 0x08
-compactList = 0x09
-compactStruct = 0x0C
+-- https://github.com/apache/parquet-format/blob/master/src/main/thrift/parquet.thrift#L875
+data EncryptionWithFooterKey = EncryptionWithFooterKey deriving (Eq, Show)
+instance Pinchable EncryptionWithFooterKey where
+    type Tag EncryptionWithFooterKey = Pinch.TStruct
+    pinch _ = Pinch.struct []
+    unpinch _ = pure EncryptionWithFooterKey
 
-toTType :: Word8 -> TType
-toTType t =
-    fromMaybe STOP $
-        M.lookup (t .&. 0x0f) $
-            M.fromList
-                [ (compactBooleanTrue, BOOL)
-                , (compactI32, I32)
-                , (compactI64, I64)
-                , (compactDouble, DOUBLE)
-                , (compactBinary, STRING)
-                , (compactList, LIST)
-                , (compactStruct, STRUCT)
-                ]
+-- https://github.com/apache/parquet-format/blob/master/src/main/thrift/parquet.thrift#L883
+data EncryptionWithColumnKey
+    = EncryptionWithColumnKey
+    { ewck_path_in_schema :: Field 1 [Text]
+    , ewck_key_metadata :: Field 2 (Maybe ByteString)
+    }
+    deriving (Eq, Show, Generic)
 
-readField ::
-    BS.ByteString -> IORef Int -> Int16 -> IO (Maybe (TType, Int16))
-readField buf pos lastFieldId = do
-    t <- readAndAdvance pos buf
-    if t .&. 0x0f == 0
-        then return Nothing
-        else do
-            let modifier = fromIntegral ((t .&. 0xf0) `shiftR` 4) :: Int16
-            identifier <-
-                if modifier == 0
-                    then readIntFromBuffer @Int16 buf pos
-                    else return (lastFieldId + modifier)
-            let elemType = toTType (t .&. 0x0f)
-            pure $ Just (elemType, identifier)
+instance Pinchable EncryptionWithColumnKey
 
-skipToStructEnd :: BS.ByteString -> IORef Int -> IO ()
-skipToStructEnd buf pos = do
-    t <- readAndAdvance pos buf
-    if t .&. 0x0f == 0
-        then return ()
-        else do
-            let modifier = fromIntegral ((t .&. 0xf0) `shiftR` 4) :: Int16
-            _identifier <-
-                if modifier == 0
-                    then readIntFromBuffer @Int16 buf pos
-                    else return 0
-            let elemType = toTType (t .&. 0x0f)
-            skipFieldData elemType buf pos
-            skipToStructEnd buf pos
+-- https://github.com/apache/parquet-format/blob/master/src/main/thrift/parquet.thrift#L893
+-- union ColumnCryptoMetaData
+data ColumnCryptoMetaData
+    = CCM_ENCRYPTION_WITH_FOOTER_KEY (Field 1 EncryptionWithFooterKey)
+    | CCM_ENCRYPTION_WITH_COLUMN_KEY (Field 2 EncryptionWithColumnKey)
+    deriving (Eq, Show, Generic)
 
-skipFieldData :: TType -> BS.ByteString -> IORef Int -> IO ()
-skipFieldData fieldType buf pos = case fieldType of
-    BOOL -> return ()
-    I32 -> void (readIntFromBuffer @Int32 buf pos)
-    I64 -> void (readIntFromBuffer @Int64 buf pos)
-    DOUBLE -> void (readIntFromBuffer @Int64 buf pos)
-    STRING -> void (readByteString buf pos)
-    LIST -> skipList buf pos
-    STRUCT -> skipToStructEnd buf pos
-    _ -> error $ "Unknown field type" ++ show fieldType
+instance Pinchable ColumnCryptoMetaData
 
-skipList :: BS.ByteString -> IORef Int -> IO ()
-skipList buf pos = do
-    sizeAndType <- readAndAdvance pos buf
-    let sizeOnly = fromIntegral ((sizeAndType `shiftR` 4) .&. 0x0f) :: Int
-    let elemType = toTType sizeAndType
-    replicateM_ sizeOnly (skipFieldData elemType buf pos)
+-- https://github.com/apache/parquet-format/blob/master/src/main/thrift/parquet.thrift#L899
+data ColumnChunk
+    = ColumnChunk
+    { cc_file_path :: Field 1 (Maybe Text)
+    , cc_file_offset :: Field 2 Int64
+    , cc_meta_data :: Field 3 (Maybe ColumnMetaData)
+    , cc_offset_index_offset :: Field 4 (Maybe Int64)
+    , cc_offset_index_length :: Field 5 (Maybe Int32)
+    , cc_column_index_offset :: Field 6 (Maybe Int64)
+    , cc_column_index_length :: Field 7 (Maybe Int32)
+    , cc_crypto_metadata :: Field 8 (Maybe ColumnCryptoMetaData)
+    , cc_encrypted_column_metadata :: Field 9 (Maybe ByteString)
+    }
+    deriving (Eq, Show, Generic)
 
-{- | This avoids reading entire bytestring at once: it uses the seekable handle
-     seeks it to the end of the file to read the metadata
--}
-readMetadataByHandleMetaSize :: FileBufferedOrSeekable -> Int -> IO FileMetadata
-readMetadataByHandleMetaSize sh metaSize = do
-    let lastFieldId = 0
-    bs <- readLastBytes (fromIntegral $ metaSize + footerSize) sh
-    bufferPos <- newIORef 0
-    readFileMetaData defaultMetadata bs bufferPos lastFieldId
+instance Pinchable ColumnChunk
 
--- | metadata starts from (L - 8 - meta_size) to L - 8 - 1.
-readMetadata :: BS.ByteString -> Int -> IO FileMetadata
-readMetadata contents size = do
-    let metadataStartPos = BS.length contents - footerSize - size
-    let metadataBytes =
-            BS.pack $
-                map (BS.index contents) [metadataStartPos .. (metadataStartPos + size - 1)]
-    let lastFieldId = 0
-    bufferPos <- newIORef (0 :: Int)
-    readFileMetaData defaultMetadata metadataBytes bufferPos lastFieldId
+-- https://github.com/apache/parquet-format/blob/master/src/main/thrift/parquet.thrift#L940
+data SortingColumn
+    = SortingColumn
+    { sc_column_idx :: Field 1 Int32
+    , sc_descending :: Field 2 Bool
+    , sc_nulls_first :: Field 3 Bool
+    }
+    deriving (Eq, Show, Generic)
 
-readFileMetaData ::
-    FileMetadata ->
-    BS.ByteString ->
-    IORef Int ->
-    Int16 ->
-    IO FileMetadata
-readFileMetaData metadata metaDataBuf bufferPos lastFieldId = do
-    fieldContents <- readField metaDataBuf bufferPos lastFieldId
-    case fieldContents of
-        Nothing -> return metadata
-        Just (_elemType, identifier) -> case identifier of
-            1 -> do
-                parsedVersion <- readIntFromBuffer @Int32 metaDataBuf bufferPos
-                readFileMetaData
-                    (metadata{version = parsedVersion})
-                    metaDataBuf
-                    bufferPos
-                    identifier
-            2 -> do
-                sizeAndType <- readAndAdvance bufferPos metaDataBuf
-                listSize <-
-                    if (sizeAndType `shiftR` 4) .&. 0x0f == 15
-                        then readVarIntFromBuffer @Int metaDataBuf bufferPos
-                        else return $ fromIntegral ((sizeAndType `shiftR` 4) .&. 0x0f)
-                let _elemType = toTType sizeAndType
-                schemaElements <-
-                    replicateM
-                        listSize
-                        (readSchemaElement defaultSchemaElement metaDataBuf bufferPos 0)
-                readFileMetaData
-                    (metadata{schema = schemaElements})
-                    metaDataBuf
-                    bufferPos
-                    identifier
-            3 -> do
-                parsedNumRows <- readIntFromBuffer @Int64 metaDataBuf bufferPos
-                readFileMetaData
-                    (metadata{numRows = fromIntegral parsedNumRows})
-                    metaDataBuf
-                    bufferPos
-                    identifier
-            4 -> do
-                sizeAndType <- readAndAdvance bufferPos metaDataBuf
-                listSize <-
-                    if (sizeAndType `shiftR` 4) .&. 0x0f == 15
-                        then readVarIntFromBuffer @Int metaDataBuf bufferPos
-                        else return $ fromIntegral ((sizeAndType `shiftR` 4) .&. 0x0f)
+instance Pinchable SortingColumn
 
-                -- TODO actually check elemType agrees (also for all the other underscored _elemType in this module)
-                let _elemType = toTType sizeAndType
-                parsedRowGroups <-
-                    replicateM listSize (readRowGroup emptyRowGroup metaDataBuf bufferPos 0)
-                readFileMetaData
-                    (metadata{rowGroups = parsedRowGroups})
-                    metaDataBuf
-                    bufferPos
-                    identifier
-            5 -> do
-                sizeAndType <- readAndAdvance bufferPos metaDataBuf
-                listSize <-
-                    if (sizeAndType `shiftR` 4) .&. 0x0f == 15
-                        then readVarIntFromBuffer @Int metaDataBuf bufferPos
-                        else return $ fromIntegral ((sizeAndType `shiftR` 4) .&. 0x0f)
+-- https://github.com/apache/parquet-format/blob/master/src/main/thrift/parquet.thrift#L958
+data RowGroup
+    = RowGroup
+    { rg_columns :: Field 1 [ColumnChunk]
+    , rg_total_byte_size :: Field 2 Int64
+    , rg_num_rows :: Field 3 Int64
+    , rg_sorting_columns :: Field 4 (Maybe [SortingColumn])
+    , rg_file_offset :: Field 5 (Maybe Int64)
+    , rg_total_compressed_size :: Field 6 (Maybe Int64)
+    , rg_ordinal :: Field 7 (Maybe Int16)
+    }
+    deriving (Eq, Show, Generic)
 
-                let _elemType = toTType sizeAndType
-                parsedKeyValueMetadata <-
-                    replicateM listSize (readKeyValue emptyKeyValue metaDataBuf bufferPos 0)
-                readFileMetaData
-                    (metadata{keyValueMetadata = parsedKeyValueMetadata})
-                    metaDataBuf
-                    bufferPos
-                    identifier
-            6 -> do
-                parsedCreatedBy <- readString metaDataBuf bufferPos
-                readFileMetaData
-                    (metadata{createdBy = Just parsedCreatedBy})
-                    metaDataBuf
-                    bufferPos
-                    identifier
-            7 -> do
-                sizeAndType <- readAndAdvance bufferPos metaDataBuf
-                listSize <-
-                    if (sizeAndType `shiftR` 4) .&. 0x0f == 15
-                        then readVarIntFromBuffer @Int metaDataBuf bufferPos
-                        else return $ fromIntegral ((sizeAndType `shiftR` 4) .&. 0x0f)
+instance Pinchable RowGroup
 
-                let _elemType = toTType sizeAndType
-                parsedColumnOrders <-
-                    replicateM listSize (readColumnOrder metaDataBuf bufferPos 0)
-                readFileMetaData
-                    (metadata{columnOrders = parsedColumnOrders})
-                    metaDataBuf
-                    bufferPos
-                    identifier
-            8 -> do
-                parsedEncryptionAlgorithm <- readEncryptionAlgorithm metaDataBuf bufferPos 0
-                readFileMetaData
-                    (metadata{encryptionAlgorithm = parsedEncryptionAlgorithm})
-                    metaDataBuf
-                    bufferPos
-                    identifier
-            9 -> do
-                parsedFooterSigningKeyMetadata <- readByteString metaDataBuf bufferPos
-                readFileMetaData
-                    (metadata{footerSigningKeyMetadata = parsedFooterSigningKeyMetadata})
-                    metaDataBuf
-                    bufferPos
-                    identifier
-            n -> return $ error $ "UNIMPLEMENTED " ++ show n
+-- https://github.com/apache/parquet-format/blob/master/src/main/thrift/parquet.thrift#L980
+data KeyValue
+    = KeyValue
+    { kv_key :: Field 1 Text
+    , kv_value :: Field 2 (Maybe Text)
+    }
+    deriving (Eq, Show, Generic)
 
-readSchemaElement ::
-    SchemaElement ->
-    BS.ByteString ->
-    IORef Int ->
-    Int16 ->
-    IO SchemaElement
-readSchemaElement schemaElement buf pos lastFieldId = do
-    fieldContents <- readField buf pos lastFieldId
-    case fieldContents of
-        Nothing -> return schemaElement
-        Just (_elemType, identifier) -> case identifier of
-            1 -> do
-                schemaElemType <- toIntegralType <$> readInt32FromBuffer buf pos
-                readSchemaElement
-                    (schemaElement{elementType = schemaElemType})
-                    buf
-                    pos
-                    identifier
-            2 -> do
-                parsedTypeLength <- readInt32FromBuffer buf pos
-                readSchemaElement
-                    (schemaElement{typeLength = parsedTypeLength})
-                    buf
-                    pos
-                    identifier
-            3 -> do
-                fieldRepetitionType <- readInt32FromBuffer buf pos
-                readSchemaElement
-                    (schemaElement{repetitionType = repetitionTypeFromInt fieldRepetitionType})
-                    buf
-                    pos
-                    identifier
-            4 -> do
-                nameSize <- readVarIntFromBuffer @Int buf pos
-                if nameSize <= 0
-                    then readSchemaElement schemaElement buf pos identifier
-                    else do
-                        contents <- replicateM nameSize (readAndAdvance pos buf)
-                        readSchemaElement
-                            (schemaElement{elementName = T.pack (map (chr . fromIntegral) contents)})
-                            buf
-                            pos
-                            identifier
-            5 -> do
-                parsedNumChildren <- readInt32FromBuffer buf pos
-                readSchemaElement
-                    (schemaElement{numChildren = parsedNumChildren})
-                    buf
-                    pos
-                    identifier
-            6 -> do
-                parsedConvertedType <- readInt32FromBuffer buf pos
-                readSchemaElement
-                    (schemaElement{convertedType = parsedConvertedType})
-                    buf
-                    pos
-                    identifier
-            7 -> do
-                parsedScale <- readInt32FromBuffer buf pos
-                readSchemaElement (schemaElement{scale = parsedScale}) buf pos identifier
-            8 -> do
-                parsedPrecision <- readInt32FromBuffer buf pos
-                readSchemaElement
-                    (schemaElement{precision = parsedPrecision})
-                    buf
-                    pos
-                    identifier
-            9 -> do
-                parsedFieldId <- readInt32FromBuffer buf pos
-                readSchemaElement
-                    (schemaElement{fieldId = parsedFieldId})
-                    buf
-                    pos
-                    identifier
-            10 -> do
-                parsedLogicalType <- readLogicalType LOGICAL_TYPE_UNKNOWN buf pos 0
-                readSchemaElement
-                    (schemaElement{logicalType = parsedLogicalType})
-                    buf
-                    pos
-                    identifier
-            n -> error ("Uknown schema element: " ++ show n)
+instance Pinchable KeyValue
 
-readRowGroup ::
-    RowGroup -> BS.ByteString -> IORef Int -> Int16 -> IO RowGroup
-readRowGroup r buf pos lastFieldId = do
-    fieldContents <- readField buf pos lastFieldId
-    case fieldContents of
-        Nothing -> return r
-        Just (_elemType, identifier) -> case identifier of
-            1 -> do
-                sizeAndType <- readAndAdvance pos buf
-                listSize <-
-                    if (sizeAndType `shiftR` 4) .&. 0x0f == 15
-                        then readVarIntFromBuffer @Int buf pos
-                        else return $ fromIntegral ((sizeAndType `shiftR` 4) .&. 0x0f)
-                let _elemType = toTType sizeAndType
-                columnChunks <-
-                    replicateM listSize (readColumnChunk emptyColumnChunk buf pos 0)
-                readRowGroup (r{rowGroupColumns = columnChunks}) buf pos identifier
-            2 -> do
-                totalBytes <- readIntFromBuffer @Int64 buf pos
-                readRowGroup (r{totalByteSize = totalBytes}) buf pos identifier
-            3 -> do
-                nRows <- readIntFromBuffer @Int64 buf pos
-                readRowGroup (r{rowGroupNumRows = nRows}) buf pos identifier
-            4 -> return r
-            5 -> do
-                offset <- readIntFromBuffer @Int64 buf pos
-                readRowGroup (r{fileOffset = offset}) buf pos identifier
-            6 -> do
-                compressedSize <- readIntFromBuffer @Int64 buf pos
-                readRowGroup
-                    (r{totalCompressedSize = compressedSize})
-                    buf
-                    pos
-                    identifier
-            7 -> do
-                parsedOrdinal <- readIntFromBuffer @Int16 buf pos
-                readRowGroup (r{ordinal = parsedOrdinal}) buf pos identifier
-            _ -> error $ "Unknown row group field: " ++ show identifier
+-- https://github.com/apache/parquet-format/blob/master/src/main/thrift/parquet.thrift#L990
+-- union ColumnOrder
+data ColumnOrder
+    = TYPE_ORDER (Field 1 TypeDefinedOrder)
+    deriving (Eq, Show, Generic)
 
-readColumnChunk ::
-    ColumnChunk -> BS.ByteString -> IORef Int -> Int16 -> IO ColumnChunk
-readColumnChunk c buf pos lastFieldId = do
-    fieldContents <- readField buf pos lastFieldId
-    case fieldContents of
-        Nothing -> return c
-        Just (_elemType, identifier) -> case identifier of
-            1 -> do
-                stringSize <- readVarIntFromBuffer @Int buf pos
-                contents <-
-                    map (chr . fromIntegral) <$> replicateM stringSize (readAndAdvance pos buf)
-                readColumnChunk
-                    (c{columnChunkFilePath = contents})
-                    buf
-                    pos
-                    identifier
-            2 -> do
-                parsedMetadataFileOffset <- readIntFromBuffer @Int64 buf pos
-                readColumnChunk
-                    (c{columnChunkMetadataFileOffset = parsedMetadataFileOffset})
-                    buf
-                    pos
-                    identifier
-            3 -> do
-                columnMetadata <- readColumnMetadata emptyColumnMetadata buf pos 0
-                readColumnChunk
-                    (c{columnMetaData = columnMetadata})
-                    buf
-                    pos
-                    identifier
-            4 -> do
-                columnOffsetIndexOffset <- readIntFromBuffer @Int64 buf pos
-                readColumnChunk
-                    (c{columnChunkOffsetIndexOffset = columnOffsetIndexOffset})
-                    buf
-                    pos
-                    identifier
-            5 -> do
-                columnOffsetIndexLength <- readInt32FromBuffer buf pos
-                readColumnChunk
-                    (c{columnChunkOffsetIndexLength = columnOffsetIndexLength})
-                    buf
-                    pos
-                    identifier
-            6 -> do
-                parsedColumnIndexOffset <- readIntFromBuffer @Int64 buf pos
-                readColumnChunk
-                    (c{columnChunkColumnIndexOffset = parsedColumnIndexOffset})
-                    buf
-                    pos
-                    identifier
-            7 -> do
-                parsedColumnIndexLength <- readInt32FromBuffer buf pos
-                readColumnChunk
-                    (c{columnChunkColumnIndexLength = parsedColumnIndexLength})
-                    buf
-                    pos
-                    identifier
-            _ -> error "Unknown column chunk"
+instance Pinchable ColumnOrder
 
-readColumnMetadata ::
-    ColumnMetaData ->
-    BS.ByteString ->
-    IORef Int ->
-    Int16 ->
-    IO ColumnMetaData
-readColumnMetadata cm buf pos lastFieldId = do
-    fieldContents <- readField buf pos lastFieldId
-    case fieldContents of
-        Nothing -> return cm
-        Just (_elemType, identifier) -> case identifier of
-            1 -> do
-                cType <- parquetTypeFromInt <$> readInt32FromBuffer buf pos
-                readColumnMetadata (cm{columnType = cType}) buf pos identifier
-            2 -> do
-                sizeAndType <- readAndAdvance pos buf
-                let sizeOnly = fromIntegral ((sizeAndType `shiftR` 4) .&. 0x0f) :: Int
-                let _elemType = toTType sizeAndType
-                encodings <- replicateM sizeOnly (readParquetEncoding buf pos 0)
-                readColumnMetadata
-                    (cm{columnEncodings = encodings})
-                    buf
-                    pos
-                    identifier
-            3 -> do
-                sizeAndType <- readAndAdvance pos buf
-                let sizeOnly = fromIntegral ((sizeAndType `shiftR` 4) .&. 0x0f) :: Int
-                let _elemType = toTType sizeAndType
-                paths <- replicateM sizeOnly (readString buf pos)
-                readColumnMetadata
-                    (cm{columnPathInSchema = paths})
-                    buf
-                    pos
-                    identifier
-            4 -> do
-                cType <- compressionCodecFromInt <$> readInt32FromBuffer buf pos
-                readColumnMetadata (cm{columnCodec = cType}) buf pos identifier
-            5 -> do
-                numValues <- readIntFromBuffer @Int64 buf pos
-                readColumnMetadata (cm{columnNumValues = numValues}) buf pos identifier
-            6 -> do
-                parsedTotalUncompressedSize <- readIntFromBuffer @Int64 buf pos
-                readColumnMetadata
-                    (cm{columnTotalUncompressedSize = parsedTotalUncompressedSize})
-                    buf
-                    pos
-                    identifier
-            7 -> do
-                parsedTotalCompressedSize <- readIntFromBuffer @Int64 buf pos
-                readColumnMetadata
-                    (cm{columnTotalCompressedSize = parsedTotalCompressedSize})
-                    buf
-                    pos
-                    identifier
-            8 -> do
-                sizeAndType <- readAndAdvance pos buf
-                let sizeOnly = fromIntegral ((sizeAndType `shiftR` 4) .&. 0x0f) :: Int
-                let _elemType = toTType sizeAndType
-                parsedKeyValueMeta <-
-                    replicateM sizeOnly (readKeyValue emptyKeyValue buf pos 0)
-                readColumnMetadata
-                    (cm{columnKeyValueMetadata = parsedKeyValueMeta})
-                    buf
-                    pos
-                    identifier
-            9 -> do
-                parsedDataPageOffset <- readIntFromBuffer @Int64 buf pos
-                readColumnMetadata
-                    (cm{columnDataPageOffset = parsedDataPageOffset})
-                    buf
-                    pos
-                    identifier
-            10 -> do
-                parsedIndexPageOffset <- readIntFromBuffer @Int64 buf pos
-                readColumnMetadata
-                    (cm{columnIndexPageOffset = parsedIndexPageOffset})
-                    buf
-                    pos
-                    identifier
-            11 -> do
-                parsedDictionaryPageOffset <- readIntFromBuffer @Int64 buf pos
-                readColumnMetadata
-                    (cm{columnDictionaryPageOffset = parsedDictionaryPageOffset})
-                    buf
-                    pos
-                    identifier
-            12 -> do
-                stats <- readStatistics emptyColumnStatistics buf pos 0
-                readColumnMetadata (cm{columnStatistics = stats}) buf pos identifier
-            13 -> do
-                sizeAndType <- readAndAdvance pos buf
-                let sizeOnly = fromIntegral ((sizeAndType `shiftR` 4) .&. 0x0f) :: Int
-                let _elemType = toTType sizeAndType
-                pageEncodingStats <-
-                    replicateM sizeOnly (readPageEncodingStats emptyPageEncodingStats buf pos 0)
-                readColumnMetadata
-                    (cm{columnEncodingStats = pageEncodingStats})
-                    buf
-                    pos
-                    identifier
-            14 -> do
-                parsedBloomFilterOffset <- readIntFromBuffer @Int64 buf pos
-                readColumnMetadata
-                    (cm{bloomFilterOffset = parsedBloomFilterOffset})
-                    buf
-                    pos
-                    identifier
-            15 -> do
-                parsedBloomFilterLength <- readInt32FromBuffer buf pos
-                readColumnMetadata
-                    (cm{bloomFilterLength = parsedBloomFilterLength})
-                    buf
-                    pos
-                    identifier
-            16 -> do
-                stats <- readSizeStatistics emptySizeStatistics buf pos 0
-                readColumnMetadata
-                    (cm{columnSizeStatistics = stats})
-                    buf
-                    pos
-                    identifier
-            17 -> return $ error "UNIMPLEMENTED"
-            _ -> error $ "Unknown column metadata " ++ show identifier
+-- Empty struct for TYPE_ORDER
+data TypeDefinedOrder = TypeDefinedOrder deriving (Eq, Show)
+instance Pinchable TypeDefinedOrder where
+    type Tag TypeDefinedOrder = Pinch.TStruct
+    pinch _ = Pinch.struct []
+    unpinch _ = pure TypeDefinedOrder
 
-readEncryptionAlgorithm ::
-    BS.ByteString -> IORef Int -> Int16 -> IO EncryptionAlgorithm
-readEncryptionAlgorithm buf pos lastFieldId = do
-    fieldContents <- readField buf pos lastFieldId
-    case fieldContents of
-        Nothing -> return ENCRYPTION_ALGORITHM_UNKNOWN
-        Just (_elemType, identifier) -> case identifier of
-            1 -> do
-                readAesGcmV1
-                    ( AesGcmV1
-                        { aadPrefix = BS.empty
-                        , aadFileUnique = BS.empty
-                        , supplyAadPrefix = False
-                        }
-                    )
-                    buf
-                    pos
-                    0
-            2 -> do
-                readAesGcmCtrV1
-                    ( AesGcmCtrV1
-                        { aadPrefix = BS.empty
-                        , aadFileUnique = BS.empty
-                        , supplyAadPrefix = False
-                        }
-                    )
-                    buf
-                    pos
-                    0
-            _n -> return ENCRYPTION_ALGORITHM_UNKNOWN
+-- https://github.com/apache/parquet-format/blob/master/src/main/thrift/parquet.thrift#L1094
+data AesGcmV1
+    = AesGcmV1
+    { aes_gcm_v1_aad_prefix :: Field 1 (Maybe ByteString)
+    , aes_gcm_v1_aad_file_unique :: Field 2 (Maybe ByteString)
+    , aes_gcm_v1_supply_aad_prefix :: Field 3 (Maybe Bool)
+    }
+    deriving (Eq, Show, Generic)
 
-readColumnOrder ::
-    BS.ByteString -> IORef Int -> Int16 -> IO ColumnOrder
-readColumnOrder buf pos lastFieldId = do
-    fieldContents <- readField buf pos lastFieldId
-    case fieldContents of
-        Nothing -> return COLUMN_ORDER_UNKNOWN
-        Just (_elemType, identifier) -> case identifier of
-            1 -> do
-                -- Read begin struct and stop since this an empty struct.
-                replicateM_ 2 (readTypeOrder buf pos 0)
-                return TYPE_ORDER
-            _ -> return COLUMN_ORDER_UNKNOWN
+instance Pinchable AesGcmV1
 
-readAesGcmCtrV1 ::
-    EncryptionAlgorithm ->
-    BS.ByteString ->
-    IORef Int ->
-    Int16 ->
-    IO EncryptionAlgorithm
-readAesGcmCtrV1 v@(AesGcmCtrV1 _aadPrefix _aadFileUnique _supplyAadPrefix) buf pos lastFieldId = do
-    fieldContents <- readField buf pos lastFieldId
-    case fieldContents of
-        Nothing -> return v
-        Just (_elemType, identifier) -> case identifier of
-            1 -> do
-                parsedAadPrefix <- readByteString buf pos
-                readAesGcmCtrV1 (v{aadPrefix = parsedAadPrefix}) buf pos identifier
-            2 -> do
-                parsedAadFileUnique <- readByteString buf pos
-                readAesGcmCtrV1
-                    (v{aadFileUnique = parsedAadFileUnique})
-                    buf
-                    pos
-                    identifier
-            3 -> do
-                parsedSupplyAadPrefix <- readAndAdvance pos buf
-                readAesGcmCtrV1
-                    (v{supplyAadPrefix = parsedSupplyAadPrefix == compactBooleanTrue})
-                    buf
-                    pos
-                    identifier
-            _ -> return ENCRYPTION_ALGORITHM_UNKNOWN
-readAesGcmCtrV1 _ _ _ _ =
-    error "readAesGcmCtrV1 called with non AesGcmCtrV1"
+-- https://github.com/apache/parquet-format/blob/master/src/main/thrift/parquet.thrift#L1107
+data AesGcmCtrV1
+    = AesGcmCtrV1
+    { aes_gcm_ctr_v1_aad_prefix :: Field 1 (Maybe ByteString)
+    , aes_gcm_ctr_v1_aad_file_unique :: Field 2 (Maybe ByteString)
+    , aes_gcm_ctr_v1_supply_aad_prefix :: Field 3 (Maybe Bool)
+    }
+    deriving (Eq, Show, Generic)
 
-readAesGcmV1 ::
-    EncryptionAlgorithm ->
-    BS.ByteString ->
-    IORef Int ->
-    Int16 ->
-    IO EncryptionAlgorithm
-readAesGcmV1 v@(AesGcmV1 _aadPrefix _aadFileUnique _supplyAadPrefix) buf pos lastFieldId = do
-    fieldContents <- readField buf pos lastFieldId
-    case fieldContents of
-        Nothing -> return v
-        Just (_elemType, identifier) -> case identifier of
-            1 -> do
-                parsedAadPrefix <- readByteString buf pos
-                readAesGcmV1 (v{aadPrefix = parsedAadPrefix}) buf pos identifier
-            2 -> do
-                parsedAadFileUnique <- readByteString buf pos
-                readAesGcmV1 (v{aadFileUnique = parsedAadFileUnique}) buf pos identifier
-            3 -> do
-                parsedSupplyAadPrefix <- readAndAdvance pos buf
-                readAesGcmV1
-                    (v{supplyAadPrefix = parsedSupplyAadPrefix == compactBooleanTrue})
-                    buf
-                    pos
-                    identifier
-            _ -> return ENCRYPTION_ALGORITHM_UNKNOWN
-readAesGcmV1 _ _ _ _ =
-    error "readAesGcmV1 called with non AesGcmV1"
+instance Pinchable AesGcmCtrV1
 
-readTypeOrder ::
-    BS.ByteString -> IORef Int -> Int16 -> IO ColumnOrder
-readTypeOrder buf pos lastFieldId = do
-    fieldContents <- readField buf pos lastFieldId
-    case fieldContents of
-        Nothing -> return TYPE_ORDER
-        Just (elemType, identifier) ->
-            if elemType == STOP
-                then return TYPE_ORDER
-                else readTypeOrder buf pos identifier
+-- https://github.com/apache/parquet-format/blob/master/src/main/thrift/parquet.thrift#L1118
+-- union EncryptionAlgorithm
+data EncryptionAlgorithm
+    = AES_GCM_V1 (Field 1 AesGcmV1)
+    | AES_GCM_CTR_V1 (Field 2 AesGcmCtrV1)
+    deriving (Eq, Show, Generic)
 
-readKeyValue ::
-    KeyValue -> BS.ByteString -> IORef Int -> Int16 -> IO KeyValue
-readKeyValue kv buf pos lastFieldId = do
-    fieldContents <- readField buf pos lastFieldId
-    case fieldContents of
-        Nothing -> return kv
-        Just (_elemType, identifier) -> case identifier of
-            1 -> do
-                k <- readString buf pos
-                readKeyValue (kv{key = k}) buf pos identifier
-            2 -> do
-                v <- readString buf pos
-                readKeyValue (kv{value = v}) buf pos identifier
-            _ -> error "Unknown kv"
+instance Pinchable EncryptionAlgorithm
 
-readPageEncodingStats ::
-    PageEncodingStats ->
-    BS.ByteString ->
-    IORef Int ->
-    Int16 ->
-    IO PageEncodingStats
-readPageEncodingStats pes buf pos lastFieldId = do
-    fieldContents <- readField buf pos lastFieldId
-    case fieldContents of
-        Nothing -> return pes
-        Just (_elemType, identifier) -> case identifier of
-            1 -> do
-                pType <- pageTypeFromInt <$> readInt32FromBuffer buf pos
-                readPageEncodingStats (pes{pageEncodingPageType = pType}) buf pos identifier
-            2 -> do
-                pEnc <- parquetEncodingFromInt <$> readInt32FromBuffer buf pos
-                readPageEncodingStats (pes{pageEncoding = pEnc}) buf pos identifier
-            3 -> do
-                encodedCount <- readInt32FromBuffer buf pos
-                readPageEncodingStats
-                    (pes{pagesWithEncoding = encodedCount})
-                    buf
-                    pos
-                    identifier
-            _ -> error "Unknown page encoding stats"
+-- https://github.com/apache/parquet-format/blob/master/src/main/thrift/parquet.thrift#L1001
+data PageLocation
+    = PageLocation
+    { pl_offset :: Field 1 Int64
+    , pl_compressed_page_size :: Field 2 Int32
+    , pl_first_row_index :: Field 3 Int64
+    }
+    deriving (Eq, Show, Generic)
 
-readParquetEncoding ::
-    BS.ByteString -> IORef Int -> Int16 -> IO ParquetEncoding
-readParquetEncoding buf pos _lastFieldId = parquetEncodingFromInt <$> readInt32FromBuffer buf pos
+instance Pinchable PageLocation
 
-readStatistics ::
-    ColumnStatistics ->
-    BS.ByteString ->
-    IORef Int ->
-    Int16 ->
-    IO ColumnStatistics
-readStatistics cs buf pos lastFieldId = do
-    fieldContents <- readField buf pos lastFieldId
-    case fieldContents of
-        Nothing -> return cs
-        Just (_elemType, identifier) -> case identifier of
-            1 -> do
-                maxInBytes <- readByteString buf pos
-                readStatistics (cs{columnMax = maxInBytes}) buf pos identifier
-            2 -> do
-                minInBytes <- readByteString buf pos
-                readStatistics (cs{columnMin = minInBytes}) buf pos identifier
-            3 -> do
-                nullCount <- readIntFromBuffer @Int64 buf pos
-                readStatistics (cs{columnNullCount = nullCount}) buf pos identifier
-            4 -> do
-                distinctCount <- readIntFromBuffer @Int64 buf pos
-                readStatistics
-                    (cs{columnDistictCount = distinctCount})
-                    buf
-                    pos
-                    identifier
-            5 -> do
-                maxInBytes <- readByteString buf pos
-                readStatistics (cs{columnMaxValue = maxInBytes}) buf pos identifier
-            6 -> do
-                minInBytes <- readByteString buf pos
-                readStatistics (cs{columnMinValue = minInBytes}) buf pos identifier
-            7 -> do
-                isMaxValueExact <- readAndAdvance pos buf
-                readStatistics
-                    (cs{isColumnMaxValueExact = isMaxValueExact == compactBooleanTrue})
-                    buf
-                    pos
-                    identifier
-            8 -> do
-                isMinValueExact <- readAndAdvance pos buf
-                readStatistics
-                    (cs{isColumnMinValueExact = isMinValueExact == compactBooleanTrue})
-                    buf
-                    pos
-                    identifier
-            _ -> error "Unknown statistics"
+-- https://github.com/apache/parquet-format/blob/master/src/main/thrift/parquet.thrift#L1017
+data OffsetIndex
+    = OffsetIndex
+    { oi_page_locations :: Field 1 [PageLocation]
+    , oi_unencoded_byte_array_data_bytes :: Field 2 (Maybe [Int64])
+    }
+    deriving (Eq, Show, Generic)
 
-readSizeStatistics ::
-    SizeStatistics ->
-    BS.ByteString ->
-    IORef Int ->
-    Int16 ->
-    IO SizeStatistics
-readSizeStatistics ss buf pos lastFieldId = do
-    fieldContents <- readField buf pos lastFieldId
-    case fieldContents of
-        Nothing -> return ss
-        Just (_elemType, identifier) -> case identifier of
-            1 -> do
-                parsedUnencodedByteArrayDataTypes <- readIntFromBuffer @Int64 buf pos
-                readSizeStatistics
-                    (ss{unencodedByteArrayDataTypes = parsedUnencodedByteArrayDataTypes})
-                    buf
-                    pos
-                    identifier
-            2 -> do
-                sizeAndType <- readAndAdvance pos buf
-                let sizeOnly = fromIntegral ((sizeAndType `shiftR` 4) .&. 0x0f) :: Int
-                let _elemType = toTType sizeAndType
-                parsedRepetitionLevelHistogram <-
-                    replicateM sizeOnly (readIntFromBuffer @Int64 buf pos)
-                readSizeStatistics
-                    (ss{repetitionLevelHistogram = parsedRepetitionLevelHistogram})
-                    buf
-                    pos
-                    identifier
-            3 -> do
-                sizeAndType <- readAndAdvance pos buf
-                let sizeOnly = fromIntegral ((sizeAndType `shiftR` 4) .&. 0x0f) :: Int
-                let _elemType = toTType sizeAndType
-                parsedDefinitionLevelHistogram <-
-                    replicateM sizeOnly (readIntFromBuffer @Int64 buf pos)
-                readSizeStatistics
-                    (ss{definitionLevelHistogram = parsedDefinitionLevelHistogram})
-                    buf
-                    pos
-                    identifier
-            _ -> error "Unknown size statistics"
+instance Pinchable OffsetIndex
 
-footerSize :: Int
-footerSize = 8
+-- https://github.com/apache/parquet-format/blob/master/src/main/thrift/parquet.thrift#L1033
+data ColumnIndex
+    = ColumnIndex
+    { ci_null_pages :: Field 1 [Bool]
+    , ci_min_values :: Field 2 [ByteString]
+    , ci_max_values :: Field 3 [ByteString]
+    , ci_boundary_order :: Field 4 BoundaryOrder
+    , ci_null_counts :: Field 5 (Maybe [Int64])
+    , ci_repetition_level_histograms :: Field 6 (Maybe [Int64])
+    , ci_definition_level_histograms :: Field 7 (Maybe [Int64])
+    }
+    deriving (Eq, Show, Generic)
 
-toIntegralType :: Int32 -> TType
-toIntegralType n
-    | n == 0 = BOOL
-    | n == 1 = I32
-    | n == 2 = I64
-    | n == 3 = I96
-    | n == 4 = FLOAT
-    | n == 5 = DOUBLE
-    | n == 6 = STRING
-    | n == 7 = STRING
-    | otherwise = error ("Unknown type in schema: " ++ show n)
+instance Pinchable ColumnIndex
 
-readLogicalType ::
-    LogicalType -> BS.ByteString -> IORef Int -> Int16 -> IO LogicalType
-readLogicalType parsedLogicalType buf pos lastFieldId = do
-    fieldContents <- readField buf pos lastFieldId
-    case fieldContents of
-        Nothing -> pure parsedLogicalType
-        Just (_elemType, identifier) -> case identifier of
-            1 -> do
-                -- This is an empty enum and is read as a field.
-                _ <- readField buf pos 0
-                readLogicalType STRING_TYPE buf pos identifier
-            2 -> do
-                _ <- readField buf pos 0
-                readLogicalType MAP_TYPE buf pos identifier
-            3 -> do
-                _ <- readField buf pos 0
-                readLogicalType LIST_TYPE buf pos identifier
-            4 -> do
-                _ <- readField buf pos 0
-                readLogicalType ENUM_TYPE buf pos identifier
-            5 -> do
-                decimal <- readDecimalType 0 0 buf pos 0
-                readLogicalType decimal buf pos identifier
-            6 -> do
-                _ <- readField buf pos 0
-                readLogicalType DATE_TYPE buf pos identifier
-            7 -> do
-                time <- readTimeType False MILLISECONDS buf pos 0
-                readLogicalType time buf pos identifier
-            8 -> do
-                timestamp <- readTimestampType False MILLISECONDS buf pos 0
-                readLogicalType timestamp buf pos identifier
-            -- Apparently reserved for interval types
-            9 -> do
-                _ <- readField buf pos 0
-                readLogicalType LOGICAL_TYPE_UNKNOWN buf pos identifier
-            10 -> do
-                intType <- readIntType 0 False buf pos 0
-                readLogicalType intType buf pos identifier
-            11 -> do
-                _ <- readField buf pos 0
-                readLogicalType LOGICAL_TYPE_UNKNOWN buf pos identifier
-            12 -> do
-                _ <- readField buf pos 0
-                readLogicalType JSON_TYPE buf pos identifier
-            13 -> do
-                _ <- readField buf pos 0
-                readLogicalType BSON_TYPE buf pos identifier
-            14 -> do
-                _ <- readField buf pos 0
-                readLogicalType UUID_TYPE buf pos identifier
-            15 -> do
-                _ <- readField buf pos 0
-                readLogicalType FLOAT16_TYPE buf pos identifier
-            16 -> error "Variant fields are unsupported"
-            17 -> error "Geometry fields are unsupported"
-            18 -> error "Geography fields are unsupported"
-            n -> error $ "Unknown logical type field: " ++ show n
+-- https://github.com/apache/parquet-format/blob/master/src/main/thrift/parquet.thrift#L1248
+data DataPageHeader
+    = DataPageHeader
+    { dph_num_values :: Field 1 Int32
+    , dph_encoding :: Field 2 Encoding
+    , dph_definition_level_encoding :: Field 3 Encoding
+    , dph_repetition_level_encoding :: Field 4 Encoding
+    , dph_statistics :: Field 5 (Maybe Statistics)
+    }
+    deriving (Eq, Show, Generic)
 
-readIntType ::
-    Int8 ->
-    Bool ->
-    BS.ByteString ->
-    IORef Int ->
-    Int16 ->
-    IO LogicalType
-readIntType parsedBitWidth parsedIntIsSigned buf pos lastFieldId = do
-    t <- readAndAdvance pos buf
-    if t .&. 0x0f == 0
-        then return (IntType parsedBitWidth parsedIntIsSigned)
-        else do
-            let modifier = fromIntegral ((t .&. 0xf0) `shiftR` 4) :: Int16
-            identifier <-
-                if modifier == 0
-                    then readIntFromBuffer @Int16 buf pos
-                    else return (lastFieldId + modifier)
+instance Pinchable DataPageHeader
 
-            case identifier of
-                1 -> do
-                    bitWidth' <- readAndAdvance pos buf
-                    readIntType (fromIntegral bitWidth') parsedIntIsSigned buf pos identifier
-                2 -> do
-                    let intIsSigned' = (t .&. 0x0f) == compactBooleanTrue
-                    readIntType parsedBitWidth intIsSigned' buf pos identifier
-                _ -> error $ "UNKNOWN field ID for IntType: " ++ show identifier
+data IndexPageHeader = IndexPageHeader deriving (Eq, Show)
+instance Pinchable IndexPageHeader where
+    type Tag IndexPageHeader = Pinch.TStruct
+    pinch _ = Pinch.struct []
+    unpinch _ = pure IndexPageHeader
 
-readDecimalType ::
-    Int32 ->
-    Int32 ->
-    BS.ByteString ->
-    IORef Int ->
-    Int16 ->
-    IO LogicalType
-readDecimalType parsedPrecision parsedScale buf pos lastFieldId = do
-    fieldContents <- readField buf pos lastFieldId
-    case fieldContents of
-        Nothing -> return (DecimalType parsedPrecision parsedScale)
-        Just (_elemType, identifier) -> case identifier of
-            1 -> do
-                scale' <- readInt32FromBuffer buf pos
-                readDecimalType parsedPrecision scale' buf pos identifier
-            2 -> do
-                precision' <- readInt32FromBuffer buf pos
-                readDecimalType precision' parsedScale buf pos identifier
-            _ -> error $ "UNKNOWN field ID for DecimalType" ++ show identifier
+data DictionaryPageHeader
+    = DictionaryPageHeader
+    { diph_num_values :: Field 1 Int32
+    , diph_encoding :: Field 2 Encoding
+    , diph_is_sorted :: Field 3 (Maybe Bool)
+    }
+    deriving (Eq, Show, Generic)
 
-readTimeType ::
-    Bool ->
-    TimeUnit ->
-    BS.ByteString ->
-    IORef Int ->
-    Int16 ->
-    IO LogicalType
-readTimeType parsedIsAdjustedToUTC parsedUnit buf pos lastFieldId = do
-    fieldContents <- readField buf pos lastFieldId
-    case fieldContents of
-        Nothing ->
-            return (TimeType{isAdjustedToUTC = parsedIsAdjustedToUTC, unit = parsedUnit})
-        Just (elemType, identifier) -> case identifier of
-            1 -> do
-                let isAdjustedToUTC' = elemType == toTType compactBooleanTrue
-                readTimeType isAdjustedToUTC' parsedUnit buf pos identifier
-            2 -> do
-                unit' <- readUnit TIME_UNIT_UNKNOWN buf pos 0
-                readTimeType parsedIsAdjustedToUTC unit' buf pos identifier
-            _ -> error $ "UNKNOWN field ID for TimeType" ++ show identifier
+instance Pinchable DictionaryPageHeader
 
-readTimestampType ::
-    Bool ->
-    TimeUnit ->
-    BS.ByteString ->
-    IORef Int ->
-    Int16 ->
-    IO LogicalType
-readTimestampType parsedIsAdjustedToUTC parsedUnit buf pos lastFieldId = do
-    fieldContents <- readField buf pos lastFieldId
-    case fieldContents of
-        Nothing ->
-            return
-                (TimestampType{isAdjustedToUTC = parsedIsAdjustedToUTC, unit = parsedUnit})
-        Just (elemType, identifier) -> case identifier of
-            1 -> do
-                let isAdjustedToUTC' = elemType == toTType compactBooleanTrue
-                readTimestampType isAdjustedToUTC' parsedUnit buf pos identifier
-            2 -> do
-                unit' <- readUnit TIME_UNIT_UNKNOWN buf pos 0
-                readTimestampType parsedIsAdjustedToUTC unit' buf pos identifier
-            _ -> error $ "UNKNOWN field ID for TimestampType " ++ show identifier
+data DataPageHeaderV2
+    = DataPageHeaderV2
+    { dph2_num_values :: Field 1 Int32
+    , dph2_num_nulls :: Field 2 Int32
+    , dph2_num_rows :: Field 3 Int32
+    , dph2_encoding :: Field 4 Encoding
+    , dph2_definition_levels_byte_length :: Field 5 Int32
+    , dph2_repetition_levels_byte_length :: Field 6 Int32
+    , dph2_is_compressed :: Field 7 (Maybe Bool)
+    , dph2_statistics :: Field 8 (Maybe Statistics)
+    }
+    deriving (Eq, Show, Generic)
 
-readUnit :: TimeUnit -> BS.ByteString -> IORef Int -> Int16 -> IO TimeUnit
-readUnit parsedUnit buf pos lastFieldId = do
-    fieldContents <- readField buf pos lastFieldId
-    case fieldContents of
-        Nothing -> return parsedUnit
-        Just (_elemType, identifier) -> case identifier of
-            1 -> do
-                _ <- readField buf pos 0
-                readUnit MILLISECONDS buf pos identifier
-            2 -> do
-                _ <- readField buf pos 0
-                readUnit MICROSECONDS buf pos identifier
-            3 -> do
-                _ <- readField buf pos 0
-                readUnit NANOSECONDS buf pos identifier
-            n -> error $ "Unknown time unit: " ++ show n
+instance Pinchable DataPageHeaderV2
+
+data PageHeader
+    = PageHeader
+    { ph_type :: Field 1 PageType
+    , ph_uncompressed_page_size :: Field 2 Int32
+    , ph_compressed_page_size :: Field 3 Int32
+    , ph_crc :: Field 4 (Maybe Int32)
+    , ph_data_page_header :: Field 5 (Maybe DataPageHeader)
+    , ph_index_page_header :: Field 6 (Maybe IndexPageHeader)
+    , ph_dictionary_page_header :: Field 7 (Maybe DictionaryPageHeader)
+    , ph_data_page_header_v2 :: Field 8 (Maybe DataPageHeaderV2)
+    }
+    deriving (Eq, Show, Generic)
+
+instance Pinchable PageHeader
+
+-- https://github.com/apache/parquet-format/blob/master/src/main/thrift/parquet.thrift#L1277
+data FileMetadata
+    = FileMetadata
+    { version :: Field 1 Int32
+    , schema :: Field 2 [SchemaElement]
+    , num_rows :: Field 3 Int64
+    , row_groups :: Field 4 [RowGroup]
+    , key_value_metadata :: Field 5 (Maybe [KeyValue])
+    , created_by :: Field 6 (Maybe Text)
+    , column_orders :: Field 7 (Maybe [ColumnOrder])
+    , encryption_algorithm :: Field 8 (Maybe EncryptionAlgorithm)
+    , footer_signing_key_metadata :: Field 9 (Maybe ByteString)
+    }
+    deriving (Eq, Show, Generic)
+
+instance Pinchable FileMetadata
+
+unField :: (KnownNat n) => Field n a -> a
+unField (Pinch.Field a) = a
